% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf.R
\name{add.cf}
\alias{add.cf}
\title{Arithmetically adds two correlation functions}
\usage{
add.cf(cf1, cf2, a = 1, b = 1)
}
\arguments{
\item{cf1, cf2}{\code{cf_orig} object.}

\item{a, b}{Numeric. Factors that multiply the correlation function before
the addition.

Since addition is associative, this operates also on the bootstrap samples
and these are thus not invalidated in the process.}
}
\value{
The value is
\deqn{a C_1 + b C_2 \,.}
}
\description{
Arithmetically adds two correlation functions
}
