% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_cf.R
\name{addStat.raw_cf}
\alias{addStat.raw_cf}
\title{Extend statistics of an existing \code{raw_cf} container}
\usage{
addStat.raw_cf(cf1, cf2)
}
\arguments{
\item{cf1}{\code{raw_cf} container with or without 'data' and 'meta' mixins}

\item{cf2}{\code{raw_cf} container with or without 'data' and 'meta' mixins}
}
\description{
Extend statistics of an existing \code{raw_cf} container
}
\details{
When either of \code{cf1} or \code{cf2} does not contain any data,
the other object is returned. (allows empty \code{raw_cf} to be extended).
If the dimensions (except for the measurements) of the \code{data} fields of the
two containers match, they are concatenated along the measurement dimension.
}
