% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_online.R
\name{analysis_online}
\alias{analysis_online}
\title{analysis_online}
\usage{
analysis_online(L, T, t1, t2, beta, kappa, mul, cg_col, evals, rundir,
  cg.ylim, type = "", csw = 0, musigma = 0, mudelta = 0, muh = 0,
  addon = "", skip = 0, rectangle = TRUE, plaquette = TRUE,
  dH = TRUE, acc = TRUE, trajtime = TRUE, omeas = TRUE,
  plotsize = 5, debug = FALSE, trajlabel = FALSE, title = FALSE,
  pl = FALSE, method = "uwerr", fit.routine = "optim",
  oldnorm = FALSE, S = 1.5, stat_skip = 0, omeas.samples = 1,
  omeas.stride = 1, omeas.avg = 1, omeas.start = 0,
  omeas.stepsize = 1, evals.start = 0, evals.stepsize = 1,
  boot.R = 1500, boot.l = 2, outname_suffix = "", verbose = FALSE)
}
\arguments{
\item{L}{integer. spatial lattice extend}

\item{T}{integer. temporal lattice extend}

\item{t1}{integer. initial time of fit range}

\item{t2}{integer. end time of fit range}

\item{beta}{numeric. inverse squared gauge coupling}

\item{kappa}{numeric. hopping parameter}

\item{mul}{numeric. light twisted mass}

\item{cg_col}{integer. column of CG iteration counts to use}

\item{evals}{??. ??}

\item{rundir}{string. run directory}

\item{cg.ylim}{numeric. y-limits for CG iteration counts}

\item{type}{string. run type}

\item{csw}{numeric. clover coefficient}

\item{musigma}{numeric. average 1+1 twisted mass}

\item{mudelta}{numeric. splitting 1+1 twisted mass}

\item{muh}{numeric. heavy twisted mass}

\item{addon}{string. addon to output filenames}

\item{skip}{integer. number of initial measurements to skip in analysis}

\item{rectangle}{boolean. If true, rectangle plaquettes are analysed}

\item{plaquette}{boolean. If true, square plaquettes are analysed}

\item{dH}{boolean. If true, delta H is analysed}

\item{acc}{boolean. If true, the acceptance rate is analysed}

\item{trajtime}{boolean. If true, the time per trajectory is analysed}

\item{omeas}{boolean. If true, offline measurements are analysed}

\item{plotsize}{numeric. size of plots being generated}

\item{debug}{boolean. provide debug information}

\item{trajlabel}{boolean or string. If not \code{FALSE}, use as trajectory labels}

\item{title}{bolean or string. If not \code{FALSE}, use as main title of plots}

\item{pl}{boolean. If set to \code{TRUE} plots will be generated}

\item{method}{string. method to compute errors, can be "uwerr" or "boot"}

\item{fit.routine}{string. minimisation routine for chisq, can be "optim"}

\item{oldnorm}{boolean. If \code{TRUE} use the old tmLQCD normalisation}

\item{S}{numeric. \code{S} parameter of \link{uwerr}}

\item{stat_skip}{integer. statistics to skip}

\item{omeas.samples}{integer. number of stochastic sample per online measurement}

\item{omeas.stride}{integer. use as strides in online measurements}

\item{omeas.avg}{integer. ??}

\item{omeas.start}{integer. ??}

\item{omeas.stepsize}{integer. ??}

\item{evals.start}{integer. ??}

\item{evals.stepsize}{integer. ??}

\item{boot.R}{integer. number of bootstrap samples}

\item{boot.l}{integer. bootstrap block size}

\item{outname_suffix}{string. suffix for output files}

\item{verbose}{boolean. If \code{TRUE}, function produces verbose output.
#'}
}
\value{
a list is returned with all the accumulated results. Moreover, a
PDF file with statistics and analytics is created
}
\description{
\code{analysis_online} is a function to analyse the online measurements
and output files of the tmLQCD software, see references.
}
\references{
K. Jansen and C. Urbach, Comput.Phys.Commun. 180 (2009) 2717-2738
}
