% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf.R
\name{avg.cbt.cf}
\alias{avg.cbt.cf}
\title{average close-by-times in a correlation function}
\usage{
avg.cbt.cf(cf)
}
\arguments{
\item{cf}{object of type \link{cf}}
}
\description{
"close-by-times" averaging replaces the value of the correlation function at t
with the "hypercubic" average with the values at the neighbouring time-slices
with weights 0.25, 0.5 and 0.25
C(t') = 0.25 C(t-1) + 0.5 C(t) + 0.25 C(t+1)
where periodic boundary conditions are assumed in shift.cf
}
