% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf.R
\name{avg.ls.cf}
\alias{avg.ls.cf}
\title{Average local-smeared and smeared-local correlators}
\usage{
avg.ls.cf(cf, cols = c(2, 3))
}
\arguments{
\item{cf}{object of type \link{cf}}

\item{cols}{columns to be averaged over}
}
\description{
averages local-smeared and smeared-local correlators in cf and adjusts
nrStypes accordingly
by default, assumes that LS and SL are in columns (T/2+1)+1:3*(T/2+1)
}
