% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_cf.R
\name{block.raw_cf}
\alias{block.raw_cf}
\title{Block average correlation function data}
\usage{
block.raw_cf(cf, block_length)
}
\arguments{
\item{cf}{\code{raw_cf} object}

\item{block_length}{Integer, number of successive measurements to average over.}
}
\description{
Block \code{block_length} sequential measurements of the correlation
function together. This occurs, for example, when multiple
stochastic noise vectors are used per measurement or multiple
source locations. Alternatively, it can also be used to
account for auto-correlations in the data. If the total numbers
of measurements is not divisible by \code{block_length}, the last
measurements are discarded.
}
