% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_ts_array.R
\name{boot_ts_array}
\alias{boot_ts_array}
\title{Copy of boot:::ts.array}
\usage{
boot_ts_array(n, n.sim, R, l, sim, endcorr)
}
\description{
The function \code{ts.array} of the \code{boot} package is not exported, yet we depend
on it. It is bad to depend on private functions of a library, so we copy it
here in the current version such that the \code{boot} package is free to change
it.
}
