\name{bootstrap.cf}
\alias{bootstrap.cf}
\title{bootstrap a set of correlation functions}
\description{
  bootstrap a set of correlation functions
}
\usage{
bootstrap.cf(cf, boot.R=400, boot.l=2, seed=1234)
}
\arguments{
  \item{cf}{
    correlation matrix obtained with a call to \code{extrac.obs}.
  }
  \item{boot.R}{
    number of bootstrap samples.
  }
  \item{boot.l}{
    block size for autocorrelation analysis
  }
  \item{seed}{
    seed for the random number generation used for boostrapping.
  }
}
\value{
}
\details{
}
\references{
}
\seealso{
}
\examples{
}
\author{}
\keyword{}
