% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hankel.R
\name{bootstrap.hankel}
\alias{bootstrap.hankel}
\title{GEVP method based on Hankel matrices. Experimental.}
\usage{
bootstrap.hankel(cf, t0, n = 2, N, id = c(1), range = c(0, 1),
  eps = 0.001)
}
\arguments{
\item{cf}{object of type \link{cf}}

\item{t0}{initial time value of the GEVP, must be in between 0 and
\code{Time/2-2}. Default is 1.}

\item{n}{Integer. Size of the submatrix Hankel matrices to generate}

\item{N}{Integer.}

\item{id}{Integer.}

\item{range}{Numeric vector. Range of eigenvalues to be considered}

\item{eps}{Numeric. Cut-off}

\item{submatrix.size}{Integer. Submatrix size to be used in build
of Hankel matrices}

\item{element.order}{Integer vector. specifies how to fit the \code{n} linearly ordered single
correlators into the correlator
matrix. \code{element.order=c(1,2,3,4)} leads to a matrix
\code{matrix(cf[element.order], nrow=2)}.
Double indexing is allowed.}
}
\description{
Alternative method to determine energy levels from correlation
matrices. A so-called Hankel matrix is generated from an input
\link{cf} object and a generalised eigenvalue problem is solved
then. This is the function to call. It will perform a bootstrap
analysis.
}
\seealso{
Other hankel: \code{\link{gevp.hankel}},
  \code{\link{hankel2cf}},
  \code{\link{hankel2effectivemass}}
}
\concept{hankel}
