% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf.R
\name{bootstrap_rw.cf}
\alias{bootstrap_rw.cf}
\title{Computes the samples for reweighted correlation function}
\usage{
bootstrap_rw.cf(cf, rw, boot.R = 400, boot.l = 2, seed = 1234,
  sim = "geom", endcorr = TRUE)
}
\arguments{
\item{cf}{\code{cf} object.}

\item{rw}{\code{rw} object.}

\item{boot.R}{Integer}

\item{boot.l}{Integer}

\item{seed}{Integer}

\item{sim}{string}

\item{endcorr}{boolean}
}
\description{
Computes the samples for reweighted correlation function
}
