% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf.R
\name{c.cf}
\alias{c.cf}
\title{Concatenate correlation function objects}
\usage{
\method{c}{cf}(...)
}
\arguments{
\item{...}{One or multiple objects of type \code{cf_orig}.}
}
\description{
Concatenate correlation function objects
}
