% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hadron-package.R
\docType{data}
\name{cA2.09.48_3pi_I3_0_A1u_1_pc}
\alias{cA2.09.48_3pi_I3_0_A1u_1_pc}
\title{A three pion correlator with significant thermal states.}
\format{
An object of class \code{list} (inherits from \code{cf}, \code{cf_meta}, \code{cf_boot}, \code{cf_principal_correlator}) of length 19.
}
\usage{
cA2.09.48_3pi_I3_0_A1u_1_pc
}
\description{
A three pion correlator with significant thermal states.
}
\keyword{datasets}
