% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_shift.R
\name{calculate.energy.shift}
\alias{calculate.energy.shift}
\title{Calculate an energy shift}
\usage{
calculate.energy.shift(psamples, msamples, n, error = sd,
  print.samples = FALSE)
}
\arguments{
\item{psamples}{energy-parameter samples vector, i.e. the first column of the
matrix t with length \code{\length{boot.R}} that is returned from the n particle
fit (n > 1) with the bootstrap.nlsfit function.}

\item{msamples}{mass samples vector, i.e. the first column of the matrix t with
length \code{\length{boot.R}} that is returned from the one particle fit with
the bootstrap.nlsfit function.}

\item{n}{integer. It denotes the number of particles for which the energy shift
is calculated.}

\item{print.samples}{option for the output. If print.samples is TRUE, the energy-
shift samples are returned. Otherwise the energy-shift values and errors are
returned.}

\item{error.}{The default error is the standard deviation, which can be applied,
i.e. in the case that msamples and psamples are bootstrap samples.}
}
\description{
Calculates the n particle energy shift \code{\Delta E_n}.
}
