% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf.R
\name{cf_jackknife}
\alias{cf_jackknife}
\title{Jackknifed CF mixin constructor}
\usage{
cf_jackknife(cf, cf0, boot.l, cf.jackknife, jackknife.se)
}
\arguments{
\item{cf}{\code{cf} object to extend.}

\item{cf0}{\emph{See above}.}

\item{boot.l}{\emph{See above}.}

\item{cf.jackknife}{List, containing jackknife samples:
\itemize{
\item \code{t}: Numeric matrix, jackknifed data sets.
\item \code{t0}: Numeric vector, copy of \code{cf0}.
\item \code{l}: Integer, copy of \code{boot.l}.
}}

\item{jackknife.se}{Numeric vector, standard error over jackknife samples.}
}
\description{
With the \code{cf_jackknife} mixin, it also must have the following fields:
}
\details{
The following fields will also be made available:
\itemize{
\item \code{jackknife.samples}: Logical, indicating whether there are jackknife samples available. This is deprecated and instead the presence of bootstrap samples should be queried with \code{inherits(cf, 'cf_jackknife')}.
}
}
\seealso{
Other cf constructors: \code{\link{cf_boot}},
  \code{\link{cf_orig}},
  \code{\link{cf_principal_correlator}},
  \code{\link{cf_subtracted}}, \code{\link{cf_weighted}},
  \code{\link{cf}}
}
