% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf.R
\name{cf_orig}
\alias{cf_orig}
\title{Original data CF mixin constructor}
\usage{
cf_orig(.cf = cf(), cf, icf = NULL)
}
\arguments{
\item{.cf}{\code{cf} object to extend. Named with a leading period just to distinguish it from the member also named \code{cf}.}

\item{cf}{Numeric matrix, original data for all observables and measurements.}

\item{icf}{Numeric matrix, imaginary part of original data. Be very careful with this as quite a few functions just ignore the imaginary part and drop it in operations.}
}
\description{
Original data CF mixin constructor
}
\seealso{
Other cf constructors: 
\code{\link{cf_boot}()},
\code{\link{cf_meta}()},
\code{\link{cf_principal_correlator}()},
\code{\link{cf_shifted}()},
\code{\link{cf_smeared}()},
\code{\link{cf_subtracted}()},
\code{\link{cf_weighted}()},
\code{\link{cf}()}
}
\concept{cf constructors}
