% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf.R
\name{cf_smeared}
\alias{cf_smeared}
\title{Smeared CF mixin constructor}
\usage{
cf_smeared(cf, scf, iscf, nrSamples, obs)
}
\arguments{
\item{scf}{Like \code{cf}, but with the smeared data.}

\item{iscf}{Like \code{icf}, but with the smeared data.}

\item{nrSamples}{TODO}

\item{obs}{TODO}
}
\description{
Smeared CF mixin constructor
}
\details{
The following fields will also be made available:
\itemize{
\item \code{smeared}: Logical, whether the correlation function has smeared data. This is deprecated and instead the presence of bootstrap samples should be queried with \code{inherits(cf, 'cf_smeared')}.
}
}
\seealso{
Other cf constructors: \code{\link{cf_boot}},
  \code{\link{cf_jackknife}}, \code{\link{cf_orig}},
  \code{\link{cf_principal_correlator}},
  \code{\link{cf_subtracted}}, \code{\link{cf_weighted}},
  \code{\link{cf}}
}
