% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf.R
\name{cf_subtracted}
\alias{cf_subtracted}
\title{Subtracted CF mixin constructor}
\usage{
cf_subtracted(.cf = cf(), subtracted.values, subtracted.ii)
}
\arguments{
\item{.cf}{\code{cf} object to extend.}

\item{subtracted.values}{Numeric matrix, TODO}

\item{subtracted.ii}{Integer vector, TODO}
}
\description{
Subtracted CF mixin constructor
}
\seealso{
Other cf constructors: \code{\link{cf_boot}},
  \code{\link{cf_meta}}, \code{\link{cf_orig}},
  \code{\link{cf_principal_correlator}},
  \code{\link{cf_shifted}}, \code{\link{cf_smeared}},
  \code{\link{cf_weighted}}, \code{\link{cf}}
}
\concept{cf constructors}
