% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf.R
\name{cf_weighted}
\alias{cf_weighted}
\title{Weighted CF mixin constructor}
\usage{
cf_weighted(cf, weighted, weight.cosh, mass1, mass2)
}
\arguments{
\item{cf}{\code{cf} object to extend.}

\item{weighted}{TODO}

\item{weight.cosh}{TODO}

\item{mass1}{TODO}

\item{mass2}{TODO}
}
\description{
Weighted CF mixin constructor
}
\seealso{
Other cf constructors: \code{\link{cf_boot}},
  \code{\link{cf_jackknife}}, \code{\link{cf_orig}},
  \code{\link{cf_principal_correlator}},
  \code{\link{cf_subtracted}}, \code{\link{cf}}
}
