% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotutils.R
\name{compute.plotlims}
\alias{compute.plotlims}
\title{compute.plotlims}
\usage{
compute.plotlims(val, logscale, cumul.dval, cumul.mdval)
}
\arguments{
\item{val}{Numeric. Value.}

\item{logscale}{Boolean.}

\item{cumul.dval}{Numeric. Cumulative error.}

\item{cumul.mdval}{Numeric. Cumulative error.}
}
\description{
Computes limits for plots
}
