% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_cf.R
\name{conj_raw_cf}
\alias{conj_raw_cf}
\title{Take the complex conjugate of a \code{raw_cf} object}
\usage{
conj_raw_cf(cf)
}
\arguments{
\item{cf}{\code{raw_cf} cotnainer with data}
}
\value{
\code{raw_cf}
}
\description{
Take the complex conjugate of a \code{raw_cf} object
}
