\name{correlatormatrix}
\alias{correlatormatrix}
\docType{data}
\title{
Sample correlator matrix
}
\description{
Sample data for a correlation function for a 24 cube times 48 lattice
QCD simulation representing a pion propagation. It is stored in form of
an object of class \code{cf}, which is derived from \code{list}.
}
\usage{data("correlatormatrix")}
\format{
  list of 7 elements:
  [1] "nrObs"       "Time"        "nrStypes"    "symmetrised" "cf"         
  [6] "icf"         "cf0"
}
\examples{
data("correlatormatrix")
}
\keyword{datasets}
