% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvc_readutils.R
\name{correlators_key_meson_3pt_local}
\alias{correlators_key_meson_3pt_local}
\title{Generate HDF5 key for CVC 'correlators' meson 3pt function with a local insertion}
\usage{
correlators_key_meson_3pt_local(fwd_flav, bwd_flav, seq_flav, src_ts, dt,
  snk_gamma, cur_gamma, src_gamma, src_p, snk_p)
}
\arguments{
\item{fwd_flav}{String, "forward" quark flavour identifier.}

\item{bwd_flav}{String, "backward" quark flavour identifier.}

\item{seq_flav}{String, "sequential" quark flavour identifier.}

\item{src_ts}{Integer, source time slice.}

\item{dt}{Integer, source-sink separation.}

\item{snk_gamma}{Integer, CVC convention gamma matrix identifier at the source.}

\item{cur_gamma}{Integer, CVC convention gamma matrix identified at the insertion.}

\item{src_gamma}{Integer, CVC convention gamma matrix identified at the sink.}

\item{src_p}{Integer vector of length 3. (x,y,z) components of the source momentum
vector in lattice units.}

\item{snk_p}{Integer vector of length 3. (x,y,z) components of the sink momentum
vector in lattice units.}
}
\description{
Generate HDF5 key for CVC 'correlators' meson 3pt function with a local insertion
}
