% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deriv_utils.R
\name{create_displ_chains}
\alias{create_displ_chains}
\title{create list of chains of displacements
Multilpe covariant displacements, when applied in order, form
a list of displacments. Each consists of a direction and a dimension.}
\usage{
create_displ_chains(max_depth, dims = c(0:3), dirs = c(0, 1))
}
\arguments{
\item{max_depth}{Positive integer, number of displacement combinations
to construct.}

\item{dims}{Integer vector, which lattice dimensions to consider. Default 0:3}

\item{dirs}{Integer vector, which displacement directions to consider. Default
forward and backward <-> c(0,1)}
}
\value{
List of data frames, each with columns 'dim' and 'dir' of 'max_depth' rows.
}
\description{
create list of chains of displacements
Multilpe covariant displacements, when applied in order, form
a list of displacments. Each consists of a direction and a dimension.
}
