% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvc_readutils.R
\name{cvc_local_loop_key}
\alias{cvc_local_loop_key}
\title{Generate HDF5 key for a momentum and spin-projected CVC loop}
\usage{
cvc_local_loop_key(loop_type, istoch, gamma, p)

cvc_local_loop_key(loop_type, istoch, gamma, p)
}
\arguments{
\item{loop_type}{String, loop type.}

\item{istoch}{Integer, index of the stochastic sample.}

\item{gamma}{Integer, CVC convention gamma matrix identifier.}

\item{p}{Integer vector of length 3, (x,y,z) components of the momentum
vector in lattice units.}

\item{i_stoch}{Integer, stochastic sample id number.}
}
\description{
Generate HDF5 key for a momentum and spin-projected CVC loop

Generate key to identify a momentum and spin-projected loop
}
