% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvc_readutils.R
\name{cvc_to_raw_cf}
\alias{cvc_to_raw_cf}
\title{Convert correlation funtion in CVC HDF5 format to 'raw_cf'}
\usage{
cvc_to_raw_cf(cf_dat)
}
\arguments{
\item{cf_dat}{Numeric vector of alternating real and imaginary parts of a
correlation function.}
}
\description{
Given a numeric vector of alternating real and imaginary parts of a
correaltion function, creates an object of class 'raw_cf' with
a single measurement, inferring \code{Time} and internal dimensions
from the passed numeric vector.
}
