% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyprus_readutils.R
\name{cyprus_read_loops}
\alias{cyprus_read_loops}
\title{read HDF5 loop files in the Cyprus CalcLoops format}
\usage{
cyprus_read_loops(selections, files, Time, nstoch, accumulated = TRUE,
  legacy_traj = TRUE, verbose = FALSE)
}
\arguments{
\item{selections}{Named list with names from the list 'Scalar', 'dOp', 'Loops'
'LpsDw', 'LpsDwCv', 'LoopsCv' specifying the requesetd loop types.
The elements of this list are in turn expected
be data frames of the form
\tabular{rrr}{
\strong{qx} \tab \strong{qy} \tab \strong{qz} \cr
0           \tab 0           \tab 1           \cr
-2          \tab 1           \tab -3          \cr
...         \tab ...         \tab ...
}
specifying the momentum combinations to be extracted for each
loop type.}

\item{files}{Vector of strings, list of HDF5 files to be processed.}

\item{Time}{Integer, time extent of the lattice.}

\item{nstoch}{Integer, number of stochastic samples to be expected in file.}

\item{accumulated}{Boolean, specifies whether the loops, as organised by stochastic sample,
are accumulated, such that, say, element \code{n} corresponds to the
sum over the first \code{n} stochastic samples. If specified as \code{TRUE},
the data is post-processed to recover the measurements for the particular
samples.}

\item{verbose}{Boolean, output I/O time per file. Requires 'tictoc' package.}

\item{legecy_traj}{Boolean. The root group for the loop data is 'conf_xxxx', where 'xxxx'
corresponds to what is passed via the 'traj' flag to CalcLoops. When
left empty, this defaults to '0004'. If this was left emtpy when
the loop files were generated, set this to \code{TRUE} and the paths
will be constructed with 'conf_0004' as their root group.}
}
\value{
Named nested list of the same length as \code{selections} containg the loop data
in the \link{raw_cf} format. Each named element corresponds to one loop
type and each element of the underlying numbered list corresponds to one momentum
combination as specified via \code{selections} for this loop type in the same order.
}
\description{
The CalcLoops code produces HDF5 files which contain
a matrix of momenta and the data for the loops (without
spin projection) organised by stochastic sample. Currently, the
reading code assumes that there is a single configuration stored per
file and the "trajectory" parameter in CalcLoops is assumed
to take its default value of '4'.
}
