% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_cf.R
\name{-.raw_cf}
\alias{-.raw_cf}
\title{add two \code{raw_cf} objects}
\usage{
\method{-}{raw_cf}(cf1, cf2)
}
\arguments{
\item{cf1}{first 'raw_cf' container to be subtracted}

\item{cf2}{second 'raw_cf' container to be subtracted}
}
\value{
\code{raw_cf} object with \code{cf$data == cf1$data - cf2$data}
}
\description{
add two \code{raw_cf} objects
}
