% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy_functions.R
\name{effectivemass}
\alias{effectivemass}
\title{effectivemass}
\usage{
effectivemass(from, to, Time, Z, pl = TRUE, S, ...)
}
\arguments{
\item{from}{integer. Fit in fitrange (from, to)}

\item{to}{integer. see from.}

\item{Time}{integer. time extent of the lattice}

\item{Z}{data}

\item{pl}{boolean. plot}

\item{S}{numeric. see \link{uwerr}}

\item{...}{additional parameters passed to \link{uwerr}}
}
\value{
Returns a \link{data.frame} with named columns \code{t}, \code{mass}, \code{dmass},
\code{ddmass}, \code{tauint} and \code{dtauint}.
}
\description{
computes the effective mass with error analysis using UWerr
}
\seealso{
\link{uwerr}
}
