% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy_functions.R
\name{effmass}
\alias{effmass}
\title{effmass}
\usage{
effmass(data, timeextent, t)
}
\arguments{
\item{data}{numeric vector. data vector of length 4}

\item{timeextent}{integer. time extent of the lattice}

\item{t}{integer. physical time at which to evaluate the cosh}
}
\value{
Returns the effective mass as a single numeric value.
}
\description{
computes the effective mass via the inverse cosh
}
