% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex-utils.R
\name{escapeLatexSpecials}
\alias{escapeLatexSpecials}
\title{Escape special LaTeX characters for use in LaTeX labels}
\usage{
escapeLatexSpecials(x)
}
\arguments{
\item{x}{String or vector of strings.}
}
\description{
Escape special LaTeX characters for use in LaTeX labels
}
