% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeTemporal.cf.R
\name{extract_mass.effectivemassfit}
\alias{extract_mass.effectivemassfit}
\title{specialisation of \link{extract_mass} to objects of type
\code{effectivemassfit}}
\usage{
\method{extract_mass}{effectivemassfit}(object)
}
\arguments{
\item{object}{Object of type \code{effectivemassfit} to extract the mass from.}
}
\value{
Numeric. The mass value.
}
\description{
specialisation of \link{extract_mass} to objects of type
\code{effectivemassfit}
}
