% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeTemporal.cf.R
\name{extract_mass.matrixfit}
\alias{extract_mass.matrixfit}
\title{specialisation of \link{extract_mass} to objects of type
\code{matrixfit}}
\usage{
\method{extract_mass}{matrixfit}(object)
}
\arguments{
\item{object}{Object of type \code{matrixfit} to extract the mass from.}
}
\value{
Numeric. The mass value.
}
\description{
specialisation of \link{extract_mass} to objects of type
\code{matrixfit}
}
