% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeTemporal.cf.R
\name{extract_mass}
\alias{extract_mass}
\title{generic function to extract a fitted mass}
\usage{
extract_mass(object)
}
\arguments{
\item{object}{Object to extract the mass from.}
}
\value{
Numeric. The mass value.
}
\description{
One of the main analysis tasks in \link{hadron} is the estimation
of energy levels or masses from correlation functions. The
corresponding analysis functions return objects, typically lists,
containing the masses or energy levels. \code{extract_mass} is a
generic function to extrac such fitted mass values.
}
