% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_energyshift.R
\name{fit.scattering.length}
\alias{fit.scattering.length}
\title{Fit the n-particle energy-shift function by Bean and Savage to provided data}
\usage{
fit.scattering.length(bsamples, y, lvec, l1, l2, n, M, ain,
  useCov = FALSE, boot.fit = TRUE, fit.method = "optim",
  autoproceed = FALSE, every, priors = list(param = c(), p = c(),
  psamples = c()), cov_fn = cov, error = sd, ...)
}
\arguments{
\item{bsamples}{bootstrap samples of \code{y}.
Must be provided as array of dimensions \code{c(boot.R, n)} with \code{n}
equals to \code{length(y)}.
These samples can be computed e.g. by means of calculate.energy.shift.}

\item{y}{the energy-shift data as a one-dimensional numerical vector to be described by
the fit function. This vector can be computed e.g. by means of calculate.energy.shift.}

\item{l1}{lower limit of the fit range.}

\item{l2}{upper limit of the fit range.}

\item{n}{number of particles and index of the calculated energy shift.}

\item{M}{mass of the particles.}

\item{ain}{initial fit parameter for the scattering length.}
}
\description{
Performs a bootstrap.nlsfit for the relativistic n-particle Bean-Savage formula.
}
