% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_energyshift.R
\docType{data}
\name{geopar}
\alias{geopar}
\title{List that constains the geometric constants I, J and K that appear in the
Bean-Savage formula for the n-particle energy-shift.}
\format{An object of class \code{list} of length 3.}
\usage{
geopar
}
\description{
List that constains the geometric constants I, J and K that appear in the
Bean-Savage formula for the n-particle energy-shift.
}
\keyword{datasets}
