% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma.R
\docType{data}
\name{gm}
\alias{gm}
\title{Array of gamma structures}
\format{An object of class \code{array} of dimension 7 x 4 x 4.}
\usage{
gm
}
\description{
Array of 4x4 complex gamma matrices in the tmLQCD chiral gamma basis,
where \eqn{\gamma^5 = \gamma^0 \gamma^1 \gamma^2 \gamma^3 = } diag(c(1,1,-1,-1))
}
\details{
The index mapping is as follows
\itemize{
\item \code{gm[1,,]} \eqn{\gamma^0}
\item \code{gm[2,,]} \eqn{\gamma^1}
\item \code{gm[3,,]} \eqn{\gamma^2}
\item \code{gm[4,,]} \eqn{\gamma^3}
\item \code{gm[5,,]} \eqn{\gamma^5}
\item \code{gm[6,,]} positive parity projector \eqn{ \frac{1}{2} (1 + \gamma^0) }
\item \code{gm[7,,]} negative parity projector \eqn{ \frac{1}{2} (1 - \gamma^0) }
}

The function \code{\link{gm_mu}} can be used to access its elements using a more
"natural" indexing.
}
\concept{
Dirac

gamma
}
\keyword{datasets}
