% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma.R
\name{gm_mu}
\alias{gm_mu}
\title{Accessor function for \code{\link{gm}}}
\usage{
gm_mu(mu)
}
\arguments{
\item{mu}{Number or string denoting
\itemize{
\item Lorentz index (0,1,2,3) for \eqn{\gamma^\mu}
\item 5 for \eqn{\gamma^5}
\item "Pp" or "Pm" for the positive and negative parity projectors respectively
}}
}
\description{
Retrieve the entries of the \code{\link[hadron]{gm}} three-index array containing various
gamma structures in a natural indexing.
}
\concept{Dirac}
\concept{gamma}
