\name{hadron-internal}
\alias{arrangeCor.vector}
\alias{arrangeCor.pion}
\alias{arrangeCor.b1}
\alias{arrangeCor.a0}
\alias{getNxNmatrix}
\alias{ChiSqr.1mass}
\alias{ChiSqr.2mass}
\alias{ChiSqr.3mass}
\alias{fitmasses.vector}
\alias{fitmasses.vector.boot}
\alias{fitmasses.pion}
\alias{fitmasses.pion.boot}
\alias{fitmasses.b1}
\alias{fitmasses.b1.boot}
\alias{fitmasses.a0}
\alias{fitmasses.a0.boot}
\alias{fitf.pion}
\alias{mean.index}
\alias{fitmpcac.pion}
\alias{effectivemass}
\alias{effmass}
\alias{effmass2}
\alias{read_nissa_textcf_kernel}
\alias{compute.qtildesq}
\alias{compute.qtildesq.contdisp}
\title{Internal Hadron Functions}
\description{
  Internal hadron functions
}
\details{
  These are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written :).
}
\keyword{internal}
