\name{hadron}
\alias{hadron}
\docType{package}
\title{The Hadron Package}
\description{
  An R implementation of fitting routines used in lattice QCD. It
  provides useful functions for extraction hadronic quantities and such
  like. 
}
\details{
  An R implementation of fitting routines used in lattice QCD. It
  provides useful functions for extraction hadronic quantities and such
  like.
  
  More to come.
}
\author{Carsten Urbach, \email{carsten.urbach@liverpool.ac.uk}}
\keyword{package}
