% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hankel.R
\name{hankel2cf}
\alias{hankel2cf}
\title{hankel2cf}
\usage{
hankel2cf(hankel, id = 1)
}
\arguments{
\item{hankel}{object as returned from \link{bootstrap.hankel}}

\item{id}{Integer. ID of the principal correlator to extract}
}
\description{
hankel2cf
}
\seealso{
input is generated via \link{bootstrap.hankel}
alternatively use \link{hankel2effectivemass}.

Other hankel: 
\code{\link{bootstrap.hankel}()},
\code{\link{gevp.hankel}()},
\code{\link{hankel2effectivemass}()}
}
\concept{hankel}
