% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hankel.R
\name{hankeldensity2effectivemass}
\alias{hankeldensity2effectivemass}
\title{hankeldensity2effectivemass}
\usage{
hankeldensity2effectivemass(hankel, range = c(0, 1), method = "median")
}
\arguments{
\item{hankel}{object as returned from \link{bootstrap.hankel}}

\item{range}{Numeric vector. Value-range for the real part of the eigenvalues.
If outside this range, the eigenvalue will be discarded}

\item{method}{Character vector. Method to be used to determine the central value
of the effective mass. Must be "median" (default) or "density"}
}
\value{
Returns an object of S3 class \code{effectivemass}.
#'
}
\description{
computes the density of all bootstrap replicates of effective masses
}
\seealso{
\link{bootstrap.effectivemass}, \link{hankel2effectivemass}
}
