% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf.R
\name{has_icf}
\alias{has_icf}
\title{Checks whether the cf object contains an imaginary part}
\usage{
has_icf(.cf)
}
\arguments{
\item{.cf}{\code{cf} object}
}
\description{
Checks whether the cf object contains an imaginary part
}
