% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_cf.R
\name{int_idx_matrix.raw_cf}
\alias{int_idx_matrix.raw_cf}
\title{Construct tensor index set for the internal degrees of freedom}
\usage{
int_idx_matrix.raw_cf(cf)
}
\arguments{
\item{cf}{\code{raw_cf} container}
}
\description{
Construct tensor index set for the internal degrees of freedom
}
