% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf.R
\name{invalidate.samples.cf}
\alias{invalidate.samples.cf}
\title{Invalidate samples}
\usage{
invalidate.samples.cf(cf)
}
\arguments{
\item{cf}{\code{cf} object.

When a correlation function is modified, any resampling should be
invalidated. We could instead also choose to properly work with the samples,
but most computations are done with the original data anyway.}
}
\description{
Invalidate samples
}
