% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf.R
\name{is.cf}
\alias{is.cf}
\title{Checks whether an object is a cf}
\usage{
is.cf(x)
}
\arguments{
\item{x}{Object, possibly of class \code{cf}.}
}
\value{
Returns TRUE if the input object is of class \code{cf}, FALSE otherwise.
}
\description{
Checks whether an object is a cf
}
