% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_cf.R
\name{is.raw_cf}
\alias{is.raw_cf}
\title{check if an object is of class \code{raw_cf}}
\usage{
is.raw_cf(x)
}
\arguments{
\item{x}{object to be checked}
}
\description{
check if an object is of class \code{raw_cf}
}
