% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw.R
\name{is_empty.rw}
\alias{is_empty.rw}
\title{Checks whether the cf object contains no data}
\usage{
is_empty.rw(.rw)
}
\arguments{
\item{.rw}{\code{rw} object.}
}
\description{
Checks whether the cf object contains no data
}
\examples{
# The empty rw object must be empty:
is_empty.rw(rw())

}
