% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf.R
\name{jackknife_cov}
\alias{jackknife_cov}
\title{Computes covariance matrix for jackknife samples. The rows containing any NA will be deleted if na.rm = TRUE.}
\usage{
jackknife_cov(x, y = NULL, na.rm = FALSE, ...)
}
\description{
Computes covariance matrix for jackknife samples. The rows containing any NA will be deleted if na.rm = TRUE.
}
