\name{jackknifeafterboot}
\alias{}
\title{jackknife-after-bootstrap analysis}
\description{
  jackknife-after-bootstrap analysis for errors of errors of correlation
  functions of class \code{cf}.
}
\usage{
correlator <- jackknifeafterboot(cor)
}
\arguments{
  \item{cor}{
    An object of class \code{cf} generated by \code{\link{bootstrap.cf}}
    with \code{sim="fixed"}. 
  }
}
\value{
}
\details{
}
\references{
}
\seealso{
}
\examples{
}
\author{Carsten Urbach \email{curbach@gmx.de}}
\keyword{bootstrap, jackknife, boostrap-after-jackknife, error of error}
