% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/looptools.R
\name{loop_2pt}
\alias{loop_2pt}
\title{compute two-point correlation function between quark loops}
\usage{
loop_2pt(loop_snk, loop_src)
}
\arguments{
\item{loop_snk}{'raw_cf' container with spin-projected quark loop
at sink (see \link{loop_spin_project})}

\item{loop_src}{'raw_cf' container with spin-projected quark loop
at source (see \link{loop_spin_project})}
}
\value{
'raw_cf' container with two-point function of these two quark loops.
In the calculation, both averaging over all source locations and
the average over all stochastic sample combinations are performed.
}
\description{
compute two-point correlation function between quark loops
}
