% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/looptools.R
\name{loop_spin_project}
\alias{loop_spin_project}
\title{spin projection of quark loop data}
\usage{
loop_spin_project(loop, gamma, stochav = FALSE, factor = as.complex(1))
}
\arguments{
\item{loop}{'raw_cf' container with loop data}

\item{gamma}{4x4 complex matrix}

\item{stochav}{Boolean, specified whether the average over stochastic samples should be
performed. This makes the projection much faster
but of course prevents the projected loop data to be
used for the construction of diagrams with multiple quark loops.}

\item{factor}{Complex scaling factor to be applied.}
}
\description{
Implements the operation
\deqn{ L = a*( \Gamma_{ik} M_{ki} ) }
to give the trace of a quark loop \eqn{M} multiplied by a gamma structure \eqn{\Gamma}
and scaled by a complex factor \eqn{a}.
}
