% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/looptools.R
\name{loop_stochav}
\alias{loop_stochav}
\title{average over stochastic samples of loop}
\usage{
loop_stochav(loop)
}
\arguments{
\item{loop}{'raw_cf' container with loop data}
}
\description{
Perform mean over the third dimension of the loop data.
}
