% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hadron-package.R
\docType{data}
\name{loopdata}
\alias{loopdata}
\title{Sample loop data}
\format{
list of 10 elements: "nrObs" "Time" "nrStypes" "symmetrised" "cf"
"icf" "scf" "iscf" "nrSamples" "obs"
}
\description{
Sample data for fermion loops for a 24 cube times 48 lattice QCD
simulation. It is stored in form of a list.
}
\examples{

data("loopdata")

}
\keyword{datasets}
