% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixfit.R
\name{make_parlist}
\alias{make_parlist}
\title{Create a parameter list for \code{matrixfit}}
\usage{
make_parlist(corr_matrix_size)
}
\arguments{
\item{corr_matrix_size}{integer. Number of correlators in the matrix. This
must be a the square of an integer.}
}
\description{
Create a parameter list for \code{matrixfit}
}
