% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momentum_utils.R
\name{mom_combinations}
\alias{mom_combinations}
\title{Generate table of momentum component combinations}
\usage{
mom_combinations(psqmax)
}
\arguments{
\item{psqmax}{Integer, maximum p^2 = px^2 + py^2 + pz^2 to be included in momentum list}
}
\description{
Generate table of momentum component combinations
}
