% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf.R
\name{mul.cf}
\alias{mul.cf}
\title{Arithmetically scale a correlator}
\usage{
mul.cf(cf, a = 1)
}
\arguments{
\item{cf}{\code{cf_orig} objects.}

\item{a}{Numeric, scaling factor.}
}
\description{
Arithmetically scale a correlator
}
