% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_cf.R
\name{mul.raw_cf}
\alias{mul.raw_cf}
\title{scale \code{raw_cf} data}
\usage{
mul.raw_cf(cf, a = 1)
}
\arguments{
\item{cf}{'raw_cf' container with data to be scaled by the factor \code{a}}

\item{a}{Numeric or complex scaling factor, although it could also be
an array of dimensions compatible with \code{cf$data}}
}
\value{
\code{raw_cf} object with \code{res$data == a*cf$data}
}
\description{
scale \code{raw_cf} data
}
