% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw.R
\name{multiply.rw}
\alias{multiply.rw}
\title{Arithmetically multiplies two reweighting factors}
\usage{
multiply.rw(rw1, rw2, nf1 = 1, nf2 = 1)
}
\arguments{
\item{rw1, rw2}{\code{rw_orig} object.}

\item{nf1, nf2}{Integer. Factors that determines the number of flavours you
have for reweighting factor 1 and 2}
}
\value{
The value is
\deqn{rw_1*nf1*rw_2*nf2) \,.}
}
\description{
Arithmetically multiplies two reweighting factors
}
