% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_matrixfit.R
\name{NParticleModel}
\alias{NParticleModel}
\title{Model for the n particle correlator fit}
\description{
n particle correlator with thermal pollution term(s)
}
\section{Super class}{
\code{\link[hadron:MatrixModel]{hadron::MatrixModel}} -> \code{NParticleModel}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{NParticleModel$new()}}
\item \href{#method-prediction}{\code{NParticleModel$prediction()}}
\item \href{#method-prediction_jacobian}{\code{NParticleModel$prediction_jacobian()}}
\item \href{#method-clone}{\code{NParticleModel$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="hadron" data-topic="MatrixModel" data-id="initial_guess">}\href{../../hadron/html/MatrixModel.html#method-initial_guess}{\code{hadron::MatrixModel$initial_guess()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hadron" data-topic="MatrixModel" data-id="post_process_par">}\href{../../hadron/html/MatrixModel.html#method-post_process_par}{\code{hadron::MatrixModel$post_process_par()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NParticleModel$new(time_extent, parind, sign_vec, ov_sign_vec)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prediction"></a>}}
\subsection{Method \code{prediction()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NParticleModel$prediction(par, x, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prediction_jacobian"></a>}}
\subsection{Method \code{prediction_jacobian()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NParticleModel$prediction_jacobian(par, x, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NParticleModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
