% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_cf.R
\name{overview_plot_raw_cf}
\alias{overview_plot_raw_cf}
\title{create convenient overview plots for a \code{raw_cf} object}
\usage{
overview_plot_raw_cf(cf, reim = "real", reim_same = FALSE,
  relerr = FALSE, tauint = FALSE, value_logplot = TRUE, title = "")
}
\arguments{
\item{reim}{Vector of strings, one of 'real', 'imag' or 'both'.}

\item{reim_same}{Boolean, whether real and imaginary parts should be plotted
on the same plot. If \code{TRUE}, then \code{reim} must
be 'both'.}

\item{relerr}{Boolean, whether a plot of the relative error per time slice
should be added.}

\item{tauint}{Boolean, whether a plot of the integrated auto-correlation time
on each time slice should be added.}

\item{value_logplot}{Boolean, whether the plot of the correlator should be
on a logarithmic vertical axis. (does not affect \code{tauint}
and \code{relerr}.}

\item{title}{Character vector, will be passed as the \code{main} argument to
\link{plotwitherror} which in turn passes it to \link{plot}.}
}
\description{
create convenient overview plots for a \code{raw_cf} object
}
