% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.nlsfit.R
\name{parametric.nlsfit.cov}
\alias{parametric.nlsfit.cov}
\title{NLS fit with parametric bootstrap and covariance}
\usage{
parametric.nlsfit.cov(fn, par.guess, boot.R, y, x, cov, ...)
}
\description{
NLS fit with parametric bootstrap and covariance
}
\seealso{
Other NLS fit functions: \code{\link{bootstrap.nlsfit}},
  \code{\link{parametric.bootstrap.cov}},
  \code{\link{parametric.bootstrap}},
  \code{\link{parametric.nlsfit}},
  \code{\link{plot.bootstrapfit}},
  \code{\link{print.bootstrapfit}},
  \code{\link{summary.bootstrapfit}}
}
