% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.nlsfit.R
\name{parametric_bootstrap}
\alias{parametric_bootstrap}
\title{Parametric bootstrap}
\usage{
parametric_bootstrap(boot_R, x, dx)
}
\arguments{
\item{boot_R}{numeric. Number of bootstrap samples to generate.}

\item{x}{numeric vector. Actual values for the data.}

\item{dx}{numeric vector of the same length as \code{x} or missing. Errors of the
values.}
}
\value{
A matrix with as many columns as there are variables in \code{x} and as many rows
as \code{boot_R}.
}
\description{
Parametric bootstrap
}
\seealso{
Other NLS fit functions: \code{\link{bootstrap.nlsfit}},
  \code{\link{parametric_bootstrap_cov}},
  \code{\link{parametric_nlsfit_cov}},
  \code{\link{parametric_nlsfit}},
  \code{\link{plot.bootstrapfit}},
  \code{\link{print.bootstrapfit}},
  \code{\link{summary.bootstrapfit}}
}
