% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.nlsfit.R
\name{parametric_bootstrap_cov}
\alias{parametric_bootstrap_cov}
\title{Parametric bootstrap with covariance}
\usage{
parametric_bootstrap_cov(boot_R, x, cov)
}
\arguments{
\item{boot_R}{numeric. Number of bootstrap samples to generate.}

\item{x}{numeric vector. Actual values for the data.}

\item{cov}{numeric matrix, square, length of \code{x} or missing. Covariance
between the various variables in the vector \code{x}.}
}
\value{
A matrix with as many columns as there are variables in \code{x} and as many rows
as \code{boot_R}.
}
\description{
Parametric bootstrap with covariance
}
\seealso{
Other NLS fit functions: \code{\link{bootstrap.nlsfit}},
  \code{\link{parametric_bootstrap}},
  \code{\link{parametric_nlsfit_cov}},
  \code{\link{parametric_nlsfit}},
  \code{\link{plot.bootstrapfit}},
  \code{\link{print.bootstrapfit}},
  \code{\link{summary.bootstrapfit}}
}
