% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.nlsfit.R
\name{parametric_nlsfit}
\alias{parametric_nlsfit}
\title{NLS fit with parametric bootstrap}
\usage{
parametric_nlsfit(fn, par.guess, boot_R, y, dy, x, dx = NULL, ...)
}
\description{
NLS fit with parametric bootstrap
}
\seealso{
Other NLS fit functions: \code{\link{bootstrap.nlsfit}},
  \code{\link{parametric_bootstrap_cov}},
  \code{\link{parametric_bootstrap}},
  \code{\link{parametric_nlsfit_cov}},
  \code{\link{plot.bootstrapfit}},
  \code{\link{print.bootstrapfit}},
  \code{\link{summary.bootstrapfit}}
}
\concept{NLS fit functions}
