% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.nlsfit.R
\name{parametric_nlsfit_cov}
\alias{parametric_nlsfit_cov}
\title{NLS fit with parametric bootstrap and covariance}
\usage{
parametric_nlsfit_cov(fn, par.guess, boot_R, y, x, cov, ...)
}
\description{
NLS fit with parametric bootstrap and covariance
}
\seealso{
Other NLS fit functions: \code{\link{bootstrap.nlsfit}},
  \code{\link{parametric_bootstrap_cov}},
  \code{\link{parametric_bootstrap}},
  \code{\link{parametric_nlsfit}},
  \code{\link{plot.bootstrapfit}},
  \code{\link{print.bootstrapfit}},
  \code{\link{summary.bootstrapfit}}
}
