% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcac.R
\name{pcacfit}
\alias{pcacfit}
\title{pcacfit}
\usage{
pcacfit(data, from, to, T2, pa = FALSE)
}
\arguments{
\item{data}{'Effectivemasses` from correlators}

\item{from}{initial value of fit range}

\item{to}{final value of fit range}

\item{T2}{Time extend}

\item{pa}{Boolean.}
}
\description{
Computes the average PCAC mass
}
