% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_matrixfit.R
\docType{data}
\name{Phi4Model}
\alias{Phi4Model}
\title{3 particle correlation function}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Phi4Model
}
\description{
3 particle correlator with thermal pollution term:
\deqn{C_3 (t) = 0.5*A_1^(3)*A_2^(3)*(exp(-E_3*(T-t))+exp(-E_3*t)) + 0.5*A_3^(3)*A_4^(3)*exp(-(2/3)*E_3*(T/2)) * (exp(-(1/3)*E_3*(T-t)) + exp(-(1/3)*E_3*t))}
}
\keyword{datasets}
