\name{plaq.sample}
\alias{plaq.sample}
\docType{data}
\title{
Sample plaquette time series
}
\description{
A time series of so-called plaquette values generated by a Markov Chain
MC process using the Hybrid Monte-Carlo algorithm. Plaquettes are the
smallest possible closed loops which can be build in lattice QCD in
discretised Euclidean space time.
}
\usage{data("plaq.sample")}
\format{
  The format is:
 num [1:6352] 0.583 0.582 0.582 0.582 0.582 ...
}
\examples{
data(plaq.sample)
plot(x=c(1:length(plaq.sample)), y=plaq.sample, type="l", xlab="t", ylab="<P>")
}
\keyword{datasets}
