% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotutils.R
\name{plot.averx}
\alias{plot.averx}
\title{Plots averx data}
\usage{
\method{plot}{averx}(x, ...)
}
\arguments{
\item{x}{\code{averx} object}

\item{...}{ignored}
}
\value{
Returns the plotted data in from of a \link{data.frame} with named
columns \code{t} (the time index), \code{averx} the values of average x and
\code{daverx} the statistical error estimate.
}
\description{
Plots averx data
}
