% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotutils.R
\name{plot.b1fit}
\alias{plot.b1fit}
\title{plot.b1fit}
\usage{
\method{plot}{b1fit}(x, ...)
}
\arguments{
\item{x}{Object of type \code{b1fit}}

\item{...}{Generic graphical parameter to be passed on to \link{plotwitherror}}
}
\description{
Generic function to plot an object of type \code{b1fit}
}
