% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotutils.R
\name{plot.cfit}
\alias{plot.cfit}
\title{plot.c1fit}
\usage{
\method{plot}{cfit}(x, ...)
}
\arguments{
\item{x}{Object of type \code{c1fit}}

\item{...}{Generic graphical parameter to be passed on to \link{plotwitherror}}
}
\description{
Generic function to plot an object of type \code{c1fit}
}
