% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosh_nlsfit.R
\name{plot.coshfit}
\alias{plot.coshfit}
\title{Plot a cosh-fit}
\usage{
\method{plot}{coshfit}(x, col.fitline = "black", plot.mass = TRUE, plot.corr = FALSE, ...)
}
\arguments{
\item{x}{An object generated by \code{fit.cosh}.}

\item{col.fitline}{Color in which the fit is visualized.}

\item{plot.mass, plot.corr}{The plot can show the fitted correlator
(\code{plot.corr}) as well as the corresponding effective mass
(\code{plot.mass}, if fitted with effMass).}

\item{...}{graphical parameters to be passed on to \link{plotwitherror}}
}
\value{
No return value.
}
\description{
Plot a cosh-fit
}
