% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectivemass.R
\name{plot.effectivemass}
\alias{plot.effectivemass}
\title{plot.effectivemass}
\usage{
\method{plot}{effectivemass}(x, ..., ref.value, col, col.fitline)
}
\arguments{
\item{x}{Object of class \code{effectivemass}}

\item{...}{Graphical parameters to be passed on to \link{plotwitherror}}

\item{ref.value}{Numeric. A reference value to be plotted as a horizontal line}

\item{col}{String. Colour of the data points.}

\item{col.fitline}{String. Colour of the fitted line.}
}
\description{
plot.effectivemass
}
