% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotutils.R
\name{plot.effmass}
\alias{plot.effmass}
\title{plot.effmass}
\usage{
\method{plot}{effmass}(x, ..., ll, lf, ff)
}
\arguments{
\item{x}{Object of class \code{effmass}}

\item{...}{Graphical parameters to be passed on.}

\item{ll}{local-local effective mass object}

\item{lf}{local-fuzzed effective mass object}

\item{ff}{fuzzed-fuzzed effective mass object}
}
\description{
plot.effmass
}
