% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gevp.R
\name{plot.gevp.amplitude}
\alias{plot.gevp.amplitude}
\title{plot.gevp.amplitude}
\usage{
\method{plot}{gevp.amplitude}(x, ...)
}
\arguments{
\item{x}{Object of type \code{gevp.amplitude}.}

\item{...}{Graphical parameters to be passed on.}
}
\description{
plot.gevp.amplitude
}
