% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeacf.R
\name{plot.hadronacf}
\alias{plot.hadronacf}
\title{plot.hadronacf}
\usage{
\method{plot}{hadronacf}(x, ..., col = "black")
}
\arguments{
\item{x}{Object of type \code{hadronacf} generated by \link{computeacf}}

\item{...}{Generic graphical parameters to be passed on}

\item{col}{String. Color to be used for the data points.}
}
\description{
generic function to plot an object of class "myGamma"
}
