% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotutils.R
\name{plot.massfit}
\alias{plot.massfit}
\title{plot.massfit}
\usage{
\method{plot}{massfit}(x, ..., xlab = "t", ylab = "m")
}
\arguments{
\item{x}{Object of type \code{massfit}}

\item{...}{Generic graphical parameter to be passed on to \link{plotwitherror}}

\item{xlab}{String. Label for x-axis}

\item{ylab}{String. Lable for y-axis}
}
\description{
Generic function to plot an object of type \code{massfit}
}
