% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotutils.R
\name{plot.ofit}
\alias{plot.ofit}
\title{plot.ofit}
\usage{
\method{plot}{ofit}(x, ...)
}
\arguments{
\item{x}{Object of type \code{ofit}}

\item{...}{Generic graphical parameter to be passed on to \link{plotwitherror}}
}
\description{
Generic function to plot an object of type \code{ofit}
}
