% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotutils.R
\name{plot.pionfit}
\alias{plot.pionfit}
\title{plot.pionfit}
\usage{
\method{plot}{pionfit}(x, ...)
}
\arguments{
\item{x}{Object of type \code{pionfit}}

\item{...}{Generic graphical parameter to be passed on to \link{plotwitherror}}
}
\description{
Generic function to plot an object of type \code{pionfit}
}
