% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_cf.R
\name{plot.raw_cf}
\alias{plot.raw_cf}
\title{Concatenate correlation function objects}
\usage{
\method{plot}{raw_cf}(cf, plot_im = FALSE, plot_im_same = FALSE,
  return_only = FALSE, ...)
}
\arguments{
\item{...}{Zero or multiple objects of type \code{raw_cf}.
Concatenate two correlation function objects}
}
\description{
Concatenate correlation function objects
}
