% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotutils.R
\name{plot.rhofit}
\alias{plot.rhofit}
\title{plot.rhofit}
\usage{
\method{plot}{rhofit}(x, ...)
}
\arguments{
\item{x}{Object of type \code{rhofit}}

\item{...}{Generic graphical parameter to be passed on to \link{plotwitherror}}
}
\description{
Generic function to plot an object of type \code{rhofit}
}
