% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{plot_eigenvalue_timeseries}
\alias{plot_eigenvalue_timeseries}
\title{plot_eigenvalue_timeseries}
\usage{
plot_eigenvalue_timeseries(dat, stat_range, ylab, plotsize, filelabel,
  titletext, pdf.filename, errorband_color = rgb(0.6, 0, 0, 0.6),
  debug = FALSE)
}
\arguments{
\item{dat}{Timeseries to analyse.}

\item{stat_range}{range of statistics to use.}

\item{ylab}{Y-axis label.}

\item{plotsize}{Width and Height of plot.}

\item{filelabel}{String. Label of the file.}

\item{titletext}{Text in the plot title.}

\item{pdf.filename}{String. PDF filename.}

\item{errorband_color}{String. Colour of the error band.}

\item{debug}{Boolean. Generate debug output.}
}
\description{
function to plot timeseries of eigenvlues, including minimum and maximum eigenvalue bands
as found in the monomial_0x.data files produced by tmLQCD
}
