% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hankel.R
\name{plot_hankel_spectrum}
\alias{plot_hankel_spectrum}
\title{plot_hankel_spectrum}
\usage{
plot_hankel_spectrum(hankel, deltat = 1, id = c(1:hankel$n))
}
\arguments{
\item{hankel}{object as returned from \link{bootstrap.hankel}}

\item{deltat}{Integer. Time shift at which to plot}

\item{id}{Integer vector. Indices of eigenvalues to be plotted.
Must be part of \code{c(1:hankel$n)}.}
}
\value{
No return value.
}
\description{
produces a scatter plot of the complex \eqn{-\log}{-log}
of the eigenvalues produced by the
\link{bootstrap.hankel} method. In addition, produces a
histogramm of all real and positive eigenvalues after computing
\eqn{-\log(ev)/\delta t}{-log(ev)/delta t} in the range
(0,1) and determines its mode.
}
\seealso{
Other hankel: 
\code{\link{bootstrap.hankel_summed}()},
\code{\link{bootstrap.hankel}()},
\code{\link{gevp.hankel_summed}()},
\code{\link{gevp.hankel}()},
\code{\link{hankel2cf}()},
\code{\link{hankel2effectivemass}()}
}
\concept{hankel}
