% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{plot_timeseries}
\alias{plot_timeseries}
\title{plot_timeseries}
\usage{
plot_timeseries(dat, ylab, plotsize, titletext, hist.by, stat_range,
  pdf.filename, name = "", xlab = "$t_\\\\mathrm{MD}$",
  hist.probs = c(0, 1), errorband_color = rgb(0.6, 0, 0, 0.6),
  type = "l", uwerr.S = 2, periodogram = FALSE, debug = FALSE,
  uw.summary = TRUE, ...)
}
\arguments{
\item{dat}{Timeseries to analyse.}

\item{ylab}{Y-axis label.}

\item{plotsize}{Width and Height of plot.}

\item{titletext}{Text in the plot title.}

\item{hist.by}{Numeric. Stepping to compute the histogram breaks.}

\item{stat_range}{range of statistics to use.}

\item{pdf.filename}{String. PDF filename.}

\item{name}{String. Timeseries name.}

\item{xlab}{X-axis label.}

\item{hist.probs}{Probablities for the histogram}

\item{errorband_color}{String. Colour of the error band.}

\item{type}{String. Plot type.}

\item{uwerr.S}{Numeric. \code{S} of the \link{uwerr} method to be used.}

\item{periodogram}{Boolean. Whether to show a periodogram.}

\item{debug}{Boolean. Generate debug output.}

\item{uw.summary}{Boolean. Generate an \link{uwerr} summary.}

\item{...}{Generic graphical parameters to be passed on.}
}
\description{
function to plot timeseries data, a corresponding histogram
and an error shading for an error analysis via uwerr
}
