% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotutils.R
\name{plothlinewitherror}
\alias{plothlinewitherror}
\title{plothlinewitherror}
\usage{
plothlinewitherror(m, dp, dm, col = c("red"), x0, x1)
}
\arguments{
\item{m}{Numeric. Mean value of the line to plot.}

\item{dp}{Numeric. Error up.}

\item{dm}{Numeric. Error down.}

\item{col}{String. Colour.}

\item{x0}{Numeric. Left value of the range of the horizontal line.}

\item{x1}{Numeric. Right value of the range of the horizontal line.}
}
\description{
plot a horizontal line with error band
}
