% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_energyshift.R
\name{prediction_function}
\alias{prediction_function}
\title{prediction_function \code{fn(par, x, ...)}. The non-linear energy-shift function to be
fitted to the data. Its first argument must be the fit parameters named \code{par}.
The second must be \code{x}, the explaining variable.}
\usage{
prediction_function(par, x, M, n, ...)
}
\description{
prediction_function \code{fn(par, x, ...)}. The non-linear energy-shift function to be
fitted to the data. Its first argument must be the fit parameters named \code{par}.
The second must be \code{x}, the explaining variable.
}
