% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_energyshift.R
\name{prediction_gradient}
\alias{prediction_gradient}
\title{Gradient of the prediction_function with respect to the scattering length.}
\usage{
prediction_gradient(par, x, M, n, ...)
}
\description{
Gradient of the prediction_function with respect to the scattering length.
}
