% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3pt.R
\name{print.averx}
\alias{print.averx}
\title{print.averx}
\usage{
\method{print}{averx}(x, ...)
}
\arguments{
\item{x}{Object of type \code{averx}}

\item{...}{Generic parameters to pass on.}
}
\description{
print.averx
}
