% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.nlsfit.R
\name{print.bootstrapfit}
\alias{print.bootstrapfit}
\title{Print a bootstrap NLS fit}
\usage{
\method{print}{bootstrapfit}(x, digits = 2, ...)
}
\arguments{
\item{x}{object returned by \code{bootstrap.nlsfit}}

\item{digits}{number of significant digits to print in summary or print.}

\item{...}{Additional parameters passed to the \code{summary.bootstrapfit} function.}
}
\description{
Print a bootstrap NLS fit
}
\seealso{
Other NLS fit functions: \code{\link{bootstrap.nlsfit}},
  \code{\link{parametric.bootstrap.cov}},
  \code{\link{parametric.bootstrap}},
  \code{\link{parametric.nlsfit.cov}},
  \code{\link{parametric.nlsfit}},
  \code{\link{plot.bootstrapfit}},
  \code{\link{summary.bootstrapfit}}
}
\concept{NLS fit functions}
