% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectivemass.R
\name{print.effectivemassfit}
\alias{print.effectivemassfit}
\title{print.effectivemassfit}
\usage{
\method{print}{effectivemassfit}(x, ..., verbose = FALSE)
}
\arguments{
\item{x}{Object of class \code{effectivemass}}

\item{...}{Additional parameters to be passed on.}

\item{verbose}{Boolean. More verbose output.}
}
\description{
print.effectivemassfit
}
