% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ofit.R
\name{print.ofit}
\alias{print.ofit}
\title{print.ofit}
\usage{
\method{print}{ofit}(x, ...)
}
\arguments{
\item{x}{Object of type \code{ofit}}

\item{...}{Generic parameters to pass on.}
}
\description{
print.ofit
}
