% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.pionfit.R
\name{print.pionfit}
\alias{print.pionfit}
\title{pionfit print}
\usage{
\method{print}{pionfit}(x, ...)
}
\arguments{
\item{x}{class pionfit. object to print}

\item{...}{additional parameters to be passed on}
}
\description{
pionfit print
}
