% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_cf.R
\name{print.raw_cf}
\alias{print.raw_cf}
\title{Print summary of data contained in \code{raw_cf} container}
\usage{
\method{print}{raw_cf}(x, ...)
}
\arguments{
\item{x}{\code{raw_cf} container with data and meta-data}

\item{...}{ignored}
}
\description{
Print summary of data contained in \code{raw_cf} container
}
