% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucleon.R
\name{proton}
\alias{proton}
\title{proton}
\usage{
proton(data, bc = "antiperiodic", twistangle = 0)
}
\arguments{
\item{data}{Proton correlator matrix}

\item{bc}{String. Boundary conditions, default 'antiperiodic'}

\item{twistangle}{Numeric. Angle in twisted boundary conditions.}
}
\description{
proton
}
