\name{pscor.sample}
\alias{pscor.sample}
\docType{data}
\title{
Sample pseudoscalar correlator
}
\description{
  Sample data for a pseudoscalar correlator for time extend T=48.
}
\usage{data("pscor.sample")}
\format{
  list of 2 elements:
  [1] "t"       "ps"
}
\examples{
data("pscor.sample")
}
\keyword{datasets}
