% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_cf.R
\name{raw_cf}
\alias{raw_cf}
\title{Container for raw correlation functions}
\usage{
raw_cf()
}
\description{
This function \code{raw_cf()} creates containers for raw correlation functions
of class \code{raw_cf}. This class is particularly designed to deal with
complex and matrix-valued correlation functions emerging in statistical
mechanics and quantum field theory simulations.
Arithmetic operations are defined for this class in
several ways, as well as concatenation and \link{is.raw_cf} and \link{as.raw_cf}.
}
\seealso{
Other raw_cf constructors: \code{\link{raw_cf_data}},
  \code{\link{raw_cf_meta}}, \code{\link{raw_cf_uwerr}}
}
\concept{raw_cf constructors}
