% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_cf.R
\name{raw_cf_meta}
\alias{raw_cf_meta}
\title{\code{raw_cf} metadata mixin constructor}
\usage{
raw_cf_meta(
  cf = raw_cf(),
  nrObs = 1,
  Time = NA,
  nrStypes = 1,
  dim = c(1, 1),
  nts = Time
)
}
\arguments{
\item{cf}{initial \link{raw_cf} object}

\item{nrObs}{Integer, number of different observables assembled in the data field of this container.}

\item{Time}{Integer, full time extent.}

\item{nrStypes}{Integer, number of smearing types.}

\item{dim}{Integer vector of "internal" dimensions for matrix-valued correlation functions.
For a scalar correlation, this should be specified as \code{c(1,1)}.}

\item{nts}{Integer, number of time separations actually stored in the data field.}
}
\description{
\code{raw_cf} metadata mixin constructor
}
\seealso{
Other raw_cf constructors: 
\code{\link{raw_cf_data}()},
\code{\link{raw_cf}()}
}
\concept{raw_cf constructors}
