% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_cf.R
\name{raw_cf_to_cf}
\alias{raw_cf_to_cf}
\title{Extract a particular internal component of a 'raw_cf' into a 'cf'}
\usage{
raw_cf_to_cf(x, component)
}
\arguments{
\item{x}{'raw_cf' container with 'raw_cf_data' and 'raw_cf_meta'}

\item{component}{Integer vector of the same length as the internal dimension
of the 'raw_cf' specifying which component should be extracted.}
}
\value{
'cf' object
}
\description{
Extract a particular internal component of a 'raw_cf' into a 'cf'
}
