% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_cf.R
\name{raw_cf_uwerr}
\alias{raw_cf_uwerr}
\title{Add Gamma-method error analysis to existing \code{raw_cf} container}
\usage{
raw_cf_uwerr(cf)
}
\arguments{
\item{cf}{Correlation function container of class 'raw_cf'}
}
\description{
Add Gamma-method error analysis to existing \code{raw_cf} container
}
\seealso{
Other raw_cf constructors: \code{\link{raw_cf_data}},
  \code{\link{raw_cf_meta}}, \code{\link{raw_cf}}
}
\concept{raw_cf constructors}
