% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readutils.R
\name{read.rw}
\alias{read.rw}
\title{reading reweighting factors from a list of gauge configurations
and random samples from ASCII files}
\usage{
read.rw(file_names_to_read, gauge_conf_list, nsamples, monomial_id)
}
\arguments{
\item{file_names_to_read}{list of filenames for the reweighting factors}

\item{gauge_conf_list}{<- a list of integers with the indices of the gauge configs}

\item{nsamples}{number of stochastic samples used for computing the reweighting factors}
}
\description{
reading reweighting factors from a list of gauge configurations
and random samples from ASCII files
}
