% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readutils.R
\name{readhlcor}
\alias{readhlcor}
\title{readhlcor}
\usage{
readhlcor(filename)
}
\arguments{
\item{filename}{String. Filename of the heavy light correlator data file. The
file is expected to have nine columns, the first four integer, the second four numeric
and the last integer valued again.}
}
\value{
Invisibly returns a \link{data.frame} object containing the file content.
}
\description{
readhlcor
}
