% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeTemporal.cf.R
\name{removeTemporal.cf}
\alias{removeTemporal.cf}
\title{Remove Thermal States by Weighting and Shifting}
\usage{
removeTemporal.cf(cf, single.cf1, single.cf2, p1 = c(0, 0, 0), p2 = c(0, 0,
  0), L, lat.disp = TRUE, weight.cosh = FALSE)
}
\arguments{
\item{cf}{Object of type \link{cf}}

\item{single.cf1}{Object of type \link{cf}}

\item{single.cf2}{Object of type \link{cf}}

\item{p1}{Numeric vector. Spatial momentum of first state}

\item{p2}{Numeric vector. Spatial momentum of second state}

\item{L}{Integer. Spatial lattice extend.}

\item{lat.disp}{Boolean. Use lattice dispersion relation instead of
continuum one}

\item{weight.cosh}{Boolean. Use cosh functional form in the
weighting procedure}
}
\value{
weighted and shifted correlation function as a \link{cf} object.
}
\description{
Remove Thermal States by Weighting and Shifting
}
