% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.nlsfit.R
\name{residual_plot}
\alias{residual_plot}
\title{residual_plot}
\usage{
residual_plot(x, ...)
}
\arguments{
\item{x}{the object to plot}

\item{...}{additional parameters to be passed on to specialised functions}
}
\value{
No return value.
}
\description{
generic residual_plot method
}
