% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw.R
\name{rw}
\alias{rw}
\title{Reweighting factor container}
\usage{
rw()
}
\description{
This function \code{rw()} creates containers for reweighting factors
of class \code{rw}. This class is particularly designed to store reweighting
factors for each gauge configuration, that can be applied on
correlation functions emerging in statistical and quantum field theory
simulations. Note that the reweighting factors acts on bare cf objects
and not on bootstrapped correlation function. Multiplication operation
is defined for this class, as well as increasing statistics and \link{is.rw}.
}
\details{
#'
}
\seealso{
Other rw constructors: 
\code{\link{rw_meta}()},
\code{\link{rw_orig}()}
}
\concept{rw constructors}
