% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw.R
\name{rw_meta}
\alias{rw_meta}
\title{rw metadata mixin constructor}
\usage{
rw_meta(.rw = rw(), conf.index)
}
\arguments{
\item{.rw}{\code{rw} object to extend.}

\item{conf.index}{list of Integers, containing the index of gauge configurations}
}
\description{
rw metadata mixin constructor
}
\seealso{
Other rw constructors: 
\code{\link{rw}()}
}
\concept{rw constructors}
