% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hadron-package.R
\docType{data}
\name{samplecf}
\alias{samplecf}
\title{Sample cf data}
\format{
The format is: List of 15 $ cf : num  521 533 532 531
561 ...  $ icf : num  521 533 532 531 561 ...  $ Time : num 48
$ nrStypes : num 1 $ nrObs : num 1 $ boot.samples : logi TRUE $
jackknife.samples: logi FALSE $ symmetrised : logi TRUE $ boot.R : num 1500
$ boot.l : num 2 $ seed : num 1442556 $ sim : chr "geom" $ cf0 : num
519 375 274 221 185 ...  $ cf.tsboot :List of 11 ..$ t0 : num  519 375
274 221 185 ...  ..$ t : num  521 518 520 519 519 ...  ..$ R :
num 1500 ..$ data : num  521 533 532 531 561 ...  ..$ seed :
int  403 624 -867935848 1692432057 -1535150298 -1438296209 912697060
1838233749 1438572626 999279531 ...  ..$ statistic:function (x) ..$ sim :
chr "geom" ..$ n.sim : int 1018 ..$ call : language tsboot(tseries = cf$cf,
statistic = function(x) return(apply(x, MARGIN = 2L, FUN = mean)) ...)  ..$
l : num 2 ..$ endcorr : logi TRUE ..- attr(\emph{, "class")= chr "boot" ..-
attr(}, "boot_type")= chr "tsboot" $ tsboot.se : num  1.001 0.615
0.572 0.537 0.499 ...  - attr(*, "class")= chr "cf" "list"
}
\description{
Sample data for a correlation function for a 24 cube times 48 lattice QCD
simulation representing a pion propagation. It is stored in form of an
object of class \code{cf}, which is derived from \code{list}.
}
\examples{

data(samplecf)
bootstrapped <- bootstrap.cf(samplecf)
plot(bootstrapped)

}
\keyword{datasets}
