% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_cf.R
\name{shift.raw_cf}
\alias{shift.raw_cf}
\title{shift a \code{raw_cf} correlation function by 'places' time-slices}
\usage{
shift.raw_cf(cf, places)
}
\arguments{
\item{cf}{\code{raw_cf} container}

\item{places}{Integer, number of time slices that the correlation function
should be shifted by. Can be positive or negative.}
}
\description{
shift a \code{raw_cf} correlation function by 'places' time-slices
}
\details{
The correlation funtion \eqn{C(t)} is shifted in time to produce:
\deqn{C'(t) = C(t+places)}
using periodic boundary conditions in time.
}
