% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_matrixfit.R
\name{SingleConstantModel}
\alias{SingleConstantModel}
\title{Model for a single correlator and a simple additive constant.}
\description{
This model is just the “single” model plus a simple additive
constant. In some way it is a specialization of the “n particles” model as
the energy difference is just set to zero.
}
