% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf.R
\name{/.cf}
\alias{/.cf}
\title{Divide two cf objects by each other measurement by measurement}
\usage{
\method{/}{cf}(cf1, cf2)
}
\arguments{
\item{cf1, cf2}{\code{cf_orig} objects.}
}
\description{
Note that no complex arithmetic is used, real and imaginary parts are
treated as seperate and indepenent, such that the real part of one
is the divided by the real part of the other and similarly for the
imaginary parts.
}
\details{
Note that this is generally only allowed on bootstrap samples and mean values,
although it makes sense in some exeptional circumstances. Don't use this
function unless you're certain that you should!
}
