% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string2error.R
\name{string2error}
\alias{string2error}
\title{string2error}
\usage{
string2error(x)
}
\arguments{
\item{x}{Input character string.}
}
\value{
a numeric vector with the first element the value and the
second the error
}
\description{
takes a string of the form "x(dx)", where dx are the error digits
and returns a numeric vector c(x, y), where y is dx as a proper
numeric value.
}
\details{
can be used in combination with \link{apply}
}
\examples{
string2error("0.35667(25)")

s <- c("0.35667(25)", "0.667(50)")
apply(array(s, dim=c(1, length(s))), 2, string2error)

}
