% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3pt.R
\name{summary.averx}
\alias{summary.averx}
\title{summary.averx}
\usage{
\method{summary}{averx}(object, ...)
}
\arguments{
\item{object}{Object of type \code{averx}}

\item{...}{Generic parameters to pass on.}
}
\description{
summary.averx
}
