% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.nlsfit.R
\name{summary.bootstrapfit}
\alias{summary.bootstrapfit}
\title{Summarize a bootstrap NLS fit}
\usage{
\method{summary}{bootstrapfit}(object, digits = 2, print.correlation = TRUE)
}
\arguments{
\item{object}{object returned by \code{bootstrap.nlsfit}}

\item{digits}{number of significant digits to print in summary or print.}
}
\description{
Summarize a bootstrap NLS fit
}
\seealso{
Other NLS fit functions: \code{\link{bootstrap.nlsfit}},
  \code{\link{parametric_bootstrap_cov}},
  \code{\link{parametric_bootstrap}},
  \code{\link{parametric_nlsfit_cov}},
  \code{\link{parametric_nlsfit}},
  \code{\link{plot.bootstrapfit}},
  \code{\link{print.bootstrapfit}}
}
