% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf.R
\name{summary.cf}
\alias{summary.cf}
\title{summary.cf}
\usage{
\method{summary}{cf}(object, ...)
}
\arguments{
\item{object}{Object of type \link{cf}}

\item{...}{Generic parameters to pass on.}
}
\description{
summary.cf
}
