% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosh_nlsfit.R
\name{summary.coshfit}
\alias{summary.coshfit}
\title{Summarize a cosh-fit}
\usage{
\method{summary}{coshfit}(object, verbose = FALSE, ...)
}
\arguments{
\item{object}{An object generated by \code{fit.cosh}.}

\item{verbose}{If set to \code{TRUE}, all fit results including the
correlation matrix of the fit parameters are showed. Otherwise only the
effective mass with error is given.}

\item{...}{additional parameters to match generic \link{summary} arguments}
}
\value{
No return value.
}
\description{
Summarize a cosh-fit
}
