% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectivemass.R
\name{summary.effectivemass}
\alias{summary.effectivemass}
\title{summary.effectivemass}
\usage{
\method{summary}{effectivemass}(object, ...)
}
\arguments{
\item{object}{Object of type \link{effectivemass}}

\item{...}{Generic parameters to pass on.}
}
\description{
summary.effectivemass
}
