% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectivemass.R
\name{summary.effectivemassfit}
\alias{summary.effectivemassfit}
\title{summary.effectivemassfit}
\usage{
\method{summary}{effectivemassfit}(object, ..., verbose = FALSE)
}
\arguments{
\item{object}{Object of type \link{cf}}

\item{...}{Generic parameters to pass on.}

\item{verbose}{More verbose output.}
}
\description{
summary.effectivemassfit
}
