% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gevp.R
\name{summary.gevp.amplitude}
\alias{summary.gevp.amplitude}
\title{summary.gevp.amplitude}
\usage{
\method{summary}{gevp.amplitude}(object, ...)
}
\arguments{
\item{object}{Object of type \code{gevp.amplitude}.}

\item{...}{Generic Parameters to be passed on.}
}
\description{
summary.gevp.amplitude
}
