% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeacf.R
\name{summary.hadronacf}
\alias{summary.hadronacf}
\title{summary.hadronacf}
\usage{
\method{summary}{hadronacf}(object, ...)
}
\arguments{
\item{object}{Object of type \code{hadronacf} generated by \link{computeacf}}

\item{...}{Generic parameters to be passed on}
}
\description{
generic function to summarise an object of class "myGamma"
}
