% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hankel.R
\name{summary.hankel_summed}
\alias{summary.hankel_summed}
\title{summary.hankel_summed}
\usage{
\method{summary}{hankel_summed}(object, ...)
}
\arguments{
\item{object}{Object of type "hankel_summed" generated
by \link{bootstrap.hankel_summed}}

\item{...}{Generic parameters to pass on.}
}
\value{
Returns invisibly a data frame with columns \code{E}, \code{dE} (energies and their standard errors)
\code{q16} and \code{q84} the 16 and 84 percent quantiles.
}
\description{
summary.hankel_summed
}
