% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixfit.R
\name{summary.matrixfit}
\alias{summary.matrixfit}
\title{summary.matrixfit}
\usage{
\method{summary}{matrixfit}(object, ...)
}
\arguments{
\item{object}{Object of type \link{matrixfit}}

\item{...}{Generic parameters to pass on.}
}
\description{
summary.matrixfit
}
