% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ofit.R
\name{summary.ofit}
\alias{summary.ofit}
\title{summary.ofit}
\usage{
\method{summary}{ofit}(object, ...)
}
\arguments{
\item{object}{Object of type \code{ofit}}

\item{...}{Generic parameters to pass on.}
}
\description{
summary.ofit
}
