% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3pt.R
\name{summary.pionff}
\alias{summary.pionff}
\title{summary.pionff}
\usage{
\method{summary}{pionff}(object, ...)
}
\arguments{
\item{object}{Object of type \code{pionff}}

\item{...}{Generic parameters to pass on.}
}
\description{
summary.pionff
}
