% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.pionfit.R
\name{summary.pionfit}
\alias{summary.pionfit}
\title{pionfit summary}
\usage{
\method{summary}{pionfit}(object, ...)
}
\arguments{
\item{object}{class pionfit. object to summarise}

\item{...}{additional parameters to be passed on}
}
\description{
pionfit summary
}
