% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_cf.R
\name{summary.raw_cf}
\alias{summary.raw_cf}
\title{Print summary of data contained in \code{raw_cf} container}
\usage{
\method{summary}{raw_cf}(object, ..., statistics = FALSE)
}
\arguments{
\item{object}{\code{raw_cf} container with data and meta-data}

\item{...}{ignored}

\item{statistics}{Boolean, return central value and error
for all components of the 'raw_cf'. This can
be slow so the default is \code{FALSE}.}
}
\description{
Print summary of data contained in \code{raw_cf} container
}
