% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UWerr.R
\name{summary.uwerr}
\alias{summary.uwerr}
\title{summary.uwerr}
\usage{
\method{summary}{uwerr}(object, ...)
}
\arguments{
\item{object}{Object of type \link{uwerr}}

\item{...}{Generic parameters to pass on.}
}
\description{
summary.uwerr
}
