% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed.R
\name{swap_seed}
\alias{swap_seed}
\title{Save random number generator state}
\usage{
swap_seed(new_seed)
}
\arguments{
\item{new_seed}{integer. The new seed that is to be set. In case this is
parameter is missing, no changes are made and the function just returns
\code{NULL}. This is useful because a function can just pass on its own \code{seed}
argument and therefore control whether the seed shall be fixed or left
as-is.}
}
\value{
The old random seed is returned if it exists. Otherwise \code{NULL}. In case that
\code{new_seed} was missing, \code{NULL} is returned.
}
\description{
Save random number generator state
}
