% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeTemporal.cf.R
\name{takeTimeDiff.cf}
\alias{takeTimeDiff.cf}
\title{Take time difference}
\usage{
takeTimeDiff.cf(cf, deltat = 1, forwardshift = FALSE)
}
\arguments{
\item{cf}{Object of type \code{cf}, a particle correlation function which shall be shifted.}

\item{forwardshift}{}
}
\description{
Performs the calculation of the shifted correlator C_shift(t) = C(t) - C(t +/- deltat).
}
