% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw.R
\name{*.cf}
\alias{*.cf}
\title{Arithmetically multiply reweighting factors}
\usage{
\method{*}{cf}(rw1, rw2)
}
\arguments{
\item{rw1, rw2}{\code{rw_orig} objects.}
}
\description{
Arithmetically multiply reweighting factors
}
