% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf.R
\name{uwerr.cf}
\alias{uwerr.cf}
\title{uwerr.cf}
\usage{
uwerr.cf(cf, absval = FALSE)
}
\arguments{
\item{cf}{Object of type \code{cf}}

\item{absval}{Boolean. Use absolute values of the data.}
}
\description{
Gamma method analysis on all time-slices in a 'cf' object
}
