% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_cf.R
\name{uwerr.raw_cf}
\alias{uwerr.raw_cf}
\title{Gamma method analysis on all time-slices in a 'raw_cf' object}
\usage{
uwerr.raw_cf(cf)
}
\arguments{
\item{cf}{Correlation function container of class 'raw_cf'}
}
\description{
Gamma method analysis on all time-slices in a 'raw_cf' object
}
