% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeTemporal.cf.R
\name{weight_shift_reweight.cf}
\alias{weight_shift_reweight.cf}
\title{Weight-shift-reweight a correlation function}
\usage{
weight_shift_reweight.cf(cf, energy_difference_val, energy_difference_boot,
  cosh_factor)
}
\arguments{
\item{cf}{cf_orig and possibly cf_boot object.}

\item{energy_difference_val}{numeric. A single energy value \eqn{\Delta E}{Delta E} for
the weighting.}

\item{energy_difference_boot}{numeric vector. Samples for the energy
difference value.}

\item{cosh_factor}{integer, either \code{+1} or \code{-1}. Determines the sign $c$ in
the weight factor.}
}
\description{
The correlation function is weighted with \code{\link{weight.cf}}, then shifted, and
then weighted again with the inverse weighting factor.
}
