% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf.R
\name{addConfIndex2cf}
\alias{addConfIndex2cf}
\title{add a configuration index to an \code{cf} object}
\usage{
addConfIndex2cf(cf, conf.index)
}
\arguments{
\item{cf}{and object of class \code{cf}}

\item{conf.index}{a configuration index of the same length as \code{cf}.}
}
\value{
Returns an object of class \code{cf} equal to the input but with
element \code{conf.index} added
}
\description{
add a configuration number index to \code{cf} object.
}
\examples{

data(samplecf)
conf.index <- c(1:1018)
samplecf <- addConfIndex2cf(samplecf, conf.index=conf.index)

}
\seealso{
\link{cf}
}
\author{
Carsten Urbach, \email{urbach@hiskp.uni-bonn.de}
}
\keyword{analysis}
\keyword{bootstrap}
\keyword{correlator}
