% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf.R
\name{addStat.cf}
\alias{addStat.cf}
\title{Combine statistics of two cf objects}
\usage{
addStat.cf(cf1, cf2)
}
\arguments{
\item{cf1}{the first of the two \code{cf} objects to be combined}

\item{cf2}{the second of the two \code{cf} objects to be combined}
}
\value{
an object of class \code{cf} with the statistics of the two input
\code{cf} objects combined
}
\description{
\code{addStat.cf} takes the raw data of two \code{cf} objects and combines
them into one
}
\details{
Note that the two \code{cf} objects to be combined need to be compatible.
Otherwise, \code{addStat.cf} will abort with an error.
}
\examples{

data(samplecf)
## the following is not useful, but
## explains the usage
cfnew <- addStat.cf(cf1=samplecf, cf2=samplecf)

}
\seealso{
\code{\link{cf}}
}
\author{
Carsten Urbach, \email{curbach@gmx.de}
}
\keyword{correlation}
\keyword{function}
