% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_ts_array.R
\name{boot_ts_array}
\alias{boot_ts_array}
\title{Copy of boot:::ts.array}
\usage{
boot_ts_array(n, n.sim, R, l, sim, endcorr)
}
\arguments{
\item{n}{integer. Length of original data}

\item{n.sim}{The length of the simulated time series.  Typically this will
be equal to the length of the original time series but there
are situations when it will be larger.  One obvious situation
is if prediction is required.  Another situation in which
‘n.sim’ is larger than the original length is if ‘tseries’ is
a residual time series from fitting some model to the
original time series. In this case, ‘n.sim’ would usually be
the length of the original time series.}

\item{R}{A positive integer giving the number of bootstrap replicates
required.}

\item{l}{If ‘sim’ is ‘"fixed"’ then ‘l’ is the fixed block length used
in generating the replicate time series.  If ‘sim’ is
‘"geom"’ then ‘l’ is the mean of the geometric distribution
used to generate the block lengths. ‘l’ should be a positive
integer less than the length of ‘tseries’.  This argument is
not required when ‘sim’ is ‘"model"’ but it is required for
all other simulation types.}

\item{sim}{The type of simulation required to generate the replicate
time series.  The possible input values are ‘"model"’ (model
based resampling), ‘"fixed"’ (block resampling with fixed
block lengths of ‘l’), ‘"geom"’ (block resampling with block
lengths having a geometric distribution with mean ‘l’) or
‘"scramble"’ (phase scrambling).}

\item{endcorr}{boolean. whether or not to apply end correction}
}
\description{
The function \code{ts.array} of the \code{boot} package is not exported, yet we depend
on it. It is bad to depend on private functions of a library, so we copy it
here in the current version such that the \code{boot} package is free to change
it.
}
