% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hankel.R
\name{bootstrap.hankel}
\alias{bootstrap.hankel}
\title{GEVP method based on Hankel matrices.}
\usage{
bootstrap.hankel(cf, t0, n = 2, N = cf$Time/2 + 1, id = c(1),
  range = c(0, 1), eps = 0.001)
}
\arguments{
\item{cf}{object of type \link{cf}}

\item{t0}{initial time value of the GEVP, must be in between 0 and
\code{Time/2-2}. Default is 1.}

\item{n}{Integer. Size of the Hankel matrices to generate}

\item{N}{Integer. Maximal time index in correlation function to be used in
Hankel matrix}

\item{id}{Integer. Vector of indices of eigenvalues to consider, \eqn{1\leq id\leq n}{1 <= id <= n}.}

\item{range}{Numeric vector. Value-range of eigenvalues to be considered}

\item{eps}{Numeric. Cut-off: if the imaginary part of the generalised
eigenvalues is larger than eps, the eigenvalue is discarded.}
}
\description{
Alternative method to determine energy levels from correlation
matrices. A so-called Hankel matrix is generated from an input
\link{cf} object and a generalised eigenvalue problem is solved
then. This is the function to call. It will perform a bootstrap
analysis.
}
\examples{

data(correlatormatrix)
correlatormatrix <- bootstrap.cf(correlatormatrix, boot.R=400, boot.l=1, seed=132435)
t0 <- 4
correlatormatrix.gevp <- bootstrap.gevp(cf=correlatormatrix, t0=t0, element.order=c(1,2,3,4))
pc1 <- gevp2cf(gevp=correlatormatrix.gevp, id=1)
pc1.hankel <- bootstrap.hankel(cf=pc1, t0=1, n=2)
hpc1 <- hankel2cf(hankel=pc1.hankel, id=1)
plot(hpc1, log="y")
heffectivemass1 <- hankel2effectivemass(hankel=pc1.hankel, id=1)

}
\seealso{
Other hankel: 
\code{\link{gevp.hankel}()},
\code{\link{hankel2cf}()},
\code{\link{hankel2effectivemass}()}
}
\concept{hankel}
