% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_cf.R
\name{c.raw_cf}
\alias{c.raw_cf}
\title{Concatenate \code{raw_cf} correlation function objects}
\usage{
\method{c}{raw_cf}(...)
}
\arguments{
\item{...}{Zero or multiple objects of type \code{raw_cf}.}
}
\value{
Returns an object of S3 class \code{raw_cf}, the concatenation of the
input objects.
}
\description{
Concatenate \code{raw_cf} correlation function objects
}
