% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf.R
\name{cf}
\alias{cf}
\title{Correlation function container}
\usage{
cf(nrObs = 1, Time = NA, nrStypes = 0, symmetrised = FALSE)
}
\arguments{
\item{nrObs}{Integer, number of different measurements contained in this correlation function. One can use \link{c.cf} to add multiple observables into one container. This is for instance needed when passing to the \link{gevp} function.}

\item{Time}{Integer, full time extent.}

\item{nrStypes}{Integer, number of smearing types.}

\item{symmetrised}{Logical, indicating whether the correlation function has been symmetrized.}
}
\description{
This function \code{cf()} creates containers for correlation functions
of class \code{cf}. This class is particularly designed to deal with
correlation functions emerging in statistical and quantum field theory
simulations. Arithmetic operations are defined for this class in
several ways, as well as concatenation and \link{is.cf} and \link{as.cf}.
}
\details{
And last but not least, these are the fields that are used somewhere in the library but we have not figured out which mixin these should belong to:
\itemize{
\item \code{smeared}: Logical, whether the correlation function has smeared data.
\item \code{scf}: Like \code{cf}, but with the smeared data.
\item \code{iscf}: Like \code{icf}, but with the smeared data.
\item \code{nrSamples}: TODO
\item \code{conf.index}: TODO
\item \code{obs}: TODO
\item \code{N}: Integer, number of measurements.
}
}
\seealso{
Other cf constructors: \code{\link{cf_boot}},
  \code{\link{cf_jackknife}}, \code{\link{cf_orig}},
  \code{\link{cf_principal_correlator}},
  \code{\link{cf_subtracted}}, \code{\link{cf_weighted}}
}
