% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf.R
\name{cf}
\alias{cf}
\title{Correlation function container}
\usage{
cf()
}
\description{
This function \code{cf()} creates containers for correlation functions
of class \code{cf}. This class is particularly designed to deal with
correlation functions emerging in statistical and quantum field theory
simulations. Arithmetic operations are defined for this class in
several ways, as well as concatenation and \link{is.cf} and \link{as.cf}.
}
\details{
And last but not least, these are the fields that are used somewhere in the library but we have not figured out which mixin these should belong to:
\itemize{
\item \code{conf.index}: TODO
\item \code{N}: Integer, number of measurements.
\item \code{blockind}: TODO
\item \code{jack.boot.se}: TODO
}
}
\seealso{
Other cf constructors: \code{\link{cf_boot}},
  \code{\link{cf_jackknife}}, \code{\link{cf_meta}},
  \code{\link{cf_orig}},
  \code{\link{cf_principal_correlator}},
  \code{\link{cf_shifted}}, \code{\link{cf_smeared}},
  \code{\link{cf_subtracted}}, \code{\link{cf_weighted}}
}
