% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf.R
\name{cf_principal_correlator}
\alias{cf_principal_correlator}
\title{Principal correlator CF mixin constructor}
\usage{
cf_principal_correlator(cf, id)
}
\arguments{
\item{cf}{\code{cf} object to extend.}

\item{id}{Integer, number of the principal correlator from the GEVP. Ascending with eigenvalue, so \code{id = 1} is the lowest state.}
}
\description{
Principal correlator CF mixin constructor
}
\seealso{
Other cf constructors: \code{\link{cf_boot}},
  \code{\link{cf_jackknife}}, \code{\link{cf_orig}},
  \code{\link{cf_shifted}}, \code{\link{cf_smeared}},
  \code{\link{cf_subtracted}}, \code{\link{cf_weighted}},
  \code{\link{cf}}
}
