% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf.R
\name{cf_weighted}
\alias{cf_weighted}
\title{Weighted CF mixin constructor}
\usage{
cf_weighted(.cf = cf(), weight.factor, weight.cosh, mass1, mass2)
}
\arguments{
\item{weight.factor}{TODO}

\item{weight.cosh}{TODO}

\item{mass1}{TODO}

\item{mass2}{TODO}

\item{cf}{\code{cf} object to extend.}
}
\description{
Weighted CF mixin constructor
}
\details{
The following fields will also be made available:
\itemize{
\item \code{weighted}: Logical, indicating whether the correlation function has been weighted. This is deprecated and instead the presence of this should be queried with \code{inherits(cf, 'cf_weighted')}.
}
}
\seealso{
Other cf constructors: \code{\link{cf_boot}},
  \code{\link{cf_meta}}, \code{\link{cf_orig}},
  \code{\link{cf_principal_correlator}},
  \code{\link{cf_shifted}}, \code{\link{cf_smeared}},
  \code{\link{cf_subtracted}}, \code{\link{cf}}
}
\concept{cf constructors}
