% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf.R
\name{concat.cf}
\alias{concat.cf}
\title{Concatenate two correlation function objects}
\usage{
concat.cf(left, right)
}
\arguments{
\item{left, right}{\code{cf} objects to concatenate.}
}
\value{
Returns an object of class \code{cf} representing the concatenation
of the two input objects of class \code{cf}.
}
\description{
Concatenate two correlation function objects
}
