% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hadron-package.R
\docType{data}
\name{correlatormatrix}
\alias{correlatormatrix}
\title{Sample correlator matrix}
\format{
list of 7 elements: "nrObs" "Time" "nrStypes" "symmetrised" "cf"
"icf" "cf0"
}
\description{
Sample data for a correlation function for a 24 cube times 48 lattice QCD
simulation representing a pion propagation. It is stored in form of an
object of class \code{cf}, which is derived from \code{list}.
}
\examples{

data("correlatormatrix")

}
\keyword{datasets}
