% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyprus_readutils.R
\name{cyprus_make_key_scalar}
\alias{cyprus_make_key_scalar}
\title{HDF5 key for Cyprus CalcLoops scalar-type loops}
\usage{
cyprus_make_key_scalar(istoch, loop_type, cid = 4)
}
\arguments{
\item{istoch}{Integer, index of the stochastic sample that the key should
be generated for.}

\item{loop_type}{String, name of loop type. Allowed values:
'Scalar', 'dOp'}

\item{cid}{Integer, configuration number, internally produced by the CalcLoops
tool via the "trajectory" input flag. The default is '4' as this is
often not used in practice.}
}
\description{
Generates an HDF5 key (full path) for the scalar
type loops from the Cyprus CalcLoops application.
}
