% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex-utils.R
\name{escapeLatexSpecials}
\alias{escapeLatexSpecials}
\title{Escape special LaTeX characters for use in LaTeX labels}
\usage{
escapeLatexSpecials(x)
}
\arguments{
\item{x}{String or vector of strings.}
}
\value{
String or vector of strings with all occurences of "#", "$", "\%",
"&", "~", "_", "^", ">", "<" replaced by escaped
counterparts which should render fine when used in a tikz plot, for
example.
}
\description{
Escape special LaTeX characters for use in LaTeX labels
}
\references{
from https://stackoverflow.com/questions/36338629/escaping-special-latex-characters-in-r
}
