% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf.R
\name{extractSingleCor.cf}
\alias{extractSingleCor.cf}
\title{extract one single correlator object as \code{cf} object from a large
\code{cf} object.}
\usage{
extractSingleCor.cf(cf, id = c(1))
}
\arguments{
\item{cf}{input object of class \code{cf}}

\item{id}{id of the correlators in \code{cf} to be extracted}
}
\value{
A \code{cf} object containing only the single correlator

Returns an object of class \code{cf} corresponding ot the \code{id}s element
in the input object \code{cf}
}
\description{
\code{cf} objects are capable of storing several correlation functions in
form of a correlation matrix. \code{extractSingleCor.cf} lets one extract
one of them.
}
\seealso{
\code{\link{cf}}
}
\author{
Carsten Urbach \email{curbach@gmx.de}
}
\keyword{bootstrap}
\keyword{correlator}
