% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hankel.R
\name{gevp.hankel}
\alias{gevp.hankel}
\title{GEVP method based on Hankel matrices.}
\usage{
gevp.hankel(cf, t0 = 1, deltat = 1, n, N, eps = 1e-04, range = c(0,
  1), submatrix.size = 1, element.order = c(1, 2, 3, 4), id = c(1),
  debug = FALSE)
}
\arguments{
\item{cf}{object of type \link{cf}}

\item{t0}{initial time value of the GEVP, must be in between 0 and
\code{Time/2-2}. Default is 1.}

\item{deltat}{Integer. Time shift to be used to build the Hankel matrix}

\item{n}{Integer. Size of the submatrix Hankel matrices to generate}

\item{N}{Integer. Maximal time index in correlation function to be used in
Hankel matrix}

\item{eps}{Numeric. Cut-off: if the imaginary part of the generalised
eigenvalues is larger than eps, the eigenvalue is discarded.}

\item{range}{Numeric vector. Value-range of eigenvalues to be considered}

\item{submatrix.size}{Integer. Submatrix size to be used in build
of Hankel matrices. Submatrix.size > 1 is experimental.}

\item{element.order}{Integer vector. specifies how to fit the \code{n} linearly ordered single
correlators into the correlator
matrix for submatrix.size > 1. \code{element.order=c(1,2,3,4)} leads to a matrix
\code{matrix(cf[element.order], nrow=2)}.
Double indexing is allowed.}

\item{id}{Integer. Vector of indices of eigenvalues to consider.}

\item{debug}{Boolean. Enable debug output.}
}
\description{
Alternative method to determine energy levels from correlation
matrices. A so-called Hankel matrix is generated from an input
\link{cf} object and a generalised eigenvalue problem is solved
then.
}
\seealso{
Other hankel: \code{\link{bootstrap.hankel}},
  \code{\link{hankel2cf}},
  \code{\link{hankel2effectivemass}}
}
\concept{hankel}
