% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5utils.R
\name{h5_get_dataset}
\alias{h5_get_dataset}
\title{get dataset from HDF5 file}
\usage{
h5_get_dataset(h5f, key)
}
\arguments{
\item{h5f}{HDF5 file opened with \code{H5Fopen}}

\item{key}{String, full path to dataset.}
}
\description{
get dataset from HDF5 file
}
