\name{hadron-internal}
\alias{arrangeCor.vector}
\alias{arrangeCor.pion}
\alias{getNxNmatrix}
\alias{ChiSqr.1mass}
\alias{ChiSqr.2mass}
\alias{ChiSqr.3mass}
\alias{fitmasses.vector}
\alias{fitmasses.vector.boot}
\alias{fitmasses.pion}
\alias{fitmasses.pion.boot}
\alias{fitf.pion}
\alias{mean.index}
\alias{fitmpcac.pion}
\title{Internal Hadron Functions}
\description{
  Internal hadron functions
}
\details{
  These are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written :).
}
\keyword{internal}
