% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hankel.R
\name{hankel2cf}
\alias{hankel2cf}
\title{hankel2cf}
\usage{
hankel2cf(hankel, id = 1)
}
\arguments{
\item{hankel}{object of type \link{hankel}}

\item{id}{Integer. ID of the principal correlator to extract}
}
\description{
hankel2cf
}
\seealso{
Other hankel: \code{\link{bootstrap.hankel}},
  \code{\link{gevp.hankel}},
  \code{\link{hankel2effectivemass}}
}
\concept{hankel}
