% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hankel.R
\name{hankel2effectivemass}
\alias{hankel2effectivemass}
\title{hankel2effectivemass}
\usage{
hankel2effectivemass(hankel, id = c(1), type = "log")
}
\arguments{
\item{hankel}{object as returned from \link{bootstrap.hankel}}

\item{id}{Integer. ID of the principal correlator to extract}

\item{type}{Character vector. Type of effective mass to use.}
}
\description{
hankel2effectivemass
}
\seealso{
input is generated via \link{bootstrap.hankel}
alternatively use \link{hankel2effectivemass}.

Other hankel: 
\code{\link{bootstrap.hankel}()},
\code{\link{gevp.hankel}()},
\code{\link{hankel2cf}()}
}
\concept{hankel}
