% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf.R
\name{has_icf}
\alias{has_icf}
\title{Checks whether the cf object contains an imaginary part}
\usage{
has_icf(.cf)
}
\arguments{
\item{.cf}{\code{cf} object}
}
\value{
Returns \code{TRUE} if the \code{.cf} object has an element \code{icf},
which is the imaginary component of the correlation function.
}
\description{
Checks whether the cf object contains an imaginary part
}
