% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv_cosh.R
\name{invcosh}
\alias{invcosh}
\title{numerically invert the cosh function for the mass}
\usage{
invcosh(ratio, timeextent, t, eps = 1e-09, maxiterations = 1000)
}
\arguments{
\item{ratio}{Numeric. The value of the ratio.}

\item{timeextent}{Integer. Time extent of the lattice.}

\item{t}{Integer. The t-value where the ratio was taken.}

\item{eps}{Numeric. Precision of the numerical solution}

\item{maxiterations}{Integer. Maximal number of iterations to be
used in the iterative solver.}
}
\value{
A single numeric value is returned corresponding to the mass.
}
\description{
numerically invert the cosh function for the mass
}
\examples{

invcosh(1.2, timeextent=24, t=12)
}
