% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf.R
\name{is_empty.cf}
\alias{is_empty.cf}
\title{Checks whether the cf object contains no data}
\usage{
is_empty.cf(.cf)
}
\description{
Checks whether the cf object contains no data
}
\examples{
# The empty cf object must be empty:
is_empty.cf(cf())

# The sample cf must not be empty:
is_empty.cf(samplecf)
}
