% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_cf.R
\name{is_empty.raw_cf}
\alias{is_empty.raw_cf}
\title{check if an obect is of class \code{raw_cf} and empty otherwise}
\usage{
is_empty.raw_cf(x)
}
\arguments{
\item{.raw_cf}{object to be checked}
}
\description{
check if an obect is of class \code{raw_cf} and empty otherwise
}
