\name{jackknifeafterboot}
\alias{jackknifeafterboot}
\title{jackknife-after-bootstrap analysis}
\description{
  jackknife-after-bootstrap (JAB) analysis for errors of errors of correlation
  functions of class \code{cf}.
}
\usage{
correlator <- jackknifeafterboot(cor)
}
\arguments{
  \item{cor}{
    An object of class \code{cf} generated by \code{\link{bootstrap.cf}}
    with \code{sim="fixed"}. 
  }
}
\value{
  Returns an object of class \code{cf} with an element
  \code{jack.boot.se}, which is the JAB estimate of standard error of
  the standard error.
}
\details{
}
\references{
  S.N. Lahiri,
  "On the jackknife-after-bootstrap method for dependent data and its consistency properties",
  Econometric Theory, 2002, Vol. 18, 79-98
}
\seealso{
  \code{\link{bootstrap.cf}}, \code{\link{cf}}, \code{\link{jackknife.cf}}
}
\examples{
}
\author{Carsten Urbach \email{curbach@gmx.de}}
\keyword{bootstrap, jackknife, boostrap-after-jackknife, error of error}
