% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/looptools.R
\name{loop_stochav}
\alias{loop_stochav}
\title{average over stochastic samples of loop}
\usage{
loop_stochav(loop, nstoch_to_avg = "all")
}
\arguments{
\item{loop}{'raw_cf' container with loop data}

\item{nstoch_to_avg}{String or integer, number of stochastic samples
to average over. Only possible string is 'all'.
If an integer is supplied, it must be at least '1'
and at most consistent with the number of stochastic
samples in \code{loop}.}
}
\description{
Perform mean over the third dimension of the loop data.
}
