% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw.R
\name{multiply.rw}
\alias{multiply.rw}
\title{Arithmetically multiplies two reweighting factors}
\usage{
multiply.rw(rw1, rw2, nf1 = 1, nf2 = 1)
}
\arguments{
\item{rw1, rw2}{\code{rw_orig} objects to be muplitplied}

\item{nf1, nf2}{Integer. Factors that determines the number of flavours we
have for reweighting factor 1 and 2, the default value is 1, because
usually we compute the reweighting factors for Q: the product of
up and down or strange and charm determinant, there are no additional
terms in the sea determinant that have to be reweighted with the same
rw factor. In case we compute the rw1 factor only for the up quark, we
for example, we have to set nf1=2 to obtain the rw factor
for the light determinant.}
}
\value{
The value is
\deqn{rw_1*nf1*rw_2*nf2) \,.}
}
\description{
Arithmetically multiplies two reweighting factors
}
