% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw.R
\name{multiply.rw}
\alias{multiply.rw}
\title{Arithmetically multiplies two reweighting factors}
\usage{
multiply.rw(rw1, rw2, nf1 = 1, nf2 = 1)
}
\arguments{
\item{rw1, rw2}{\code{rw_orig} objects to be muplitplied}

\item{nf1, nf2}{Integer. It is greater than one, when
we want to apply for more flavours than  we have
actually computed. For example in case of nf=4
degenerate light quarks, we compute the reweighting
factor for two using Q=D*D^dagger, and we actually want
to apply it for all the four flavours, so we have to
choose in this case nf1=2,nf2=2
Note that in the process of multiplication we do not take
care of the normalization of the reweighting objects, since
it will drop out, when applying it to a correlation function.}
}
\value{
The value is
\deqn{rw_1*nf1*rw_2*nf2) \,.}
}
\description{
Arithmetically multiplies two reweighting factors
}
