% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_matrixfit.R
\docType{data}
\name{NParticleModel}
\alias{NParticleModel}
\title{Model for the n particle correlator fit}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
NParticleModel
}
\description{
n particle correlator with thermal pollution term(s)
}
\keyword{datasets}
