% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixfit.R
\name{pcModel}
\alias{pcModel}
\title{Principal correlator two state model.}
\usage{
pcModel(par, t, Time, delta1 = 1, reference_time)
}
\arguments{
\item{par}{Numeric vector: Fit parameters of the model. In an
object of type \code{matrixfit}, this should be located at
\code{$opt.res$par}.}

\item{t}{Numeric vector: Time of interest.}

\item{Time}{Numeric: Time extent of the lattice.}

\item{delta1}{dummy parameter for compatibility}

\item{reference_time}{Numeric: GEVP reference time value in physical time convention}
}
\value{
Returns a numeric vector with the same length as the input vector \code{t}
containing the model evaluation for these t-values.
}
\description{
Principal correlator two state model.
}
\seealso{
\code{\link{matrixfit}}
}
