% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixfit.R
\name{pcModel}
\alias{pcModel}
\title{Principal correlator two state model.}
\usage{
pcModel(par, t, T, delta1 = 1, reference_time)
}
\arguments{
\item{par}{Numeric vector: Fit parameters of the model. In an
object of type \code{matrixfit}, this should be located at
\code{$opt.res$par}.}

\item{t}{Numeric vector: Time of interest.}

\item{T}{Numeric: Time extent of the lattice.}

\item{reference_time}{Numeric: GEVP reference time value in physical time convention}

\item{sign.vec}{Numeric vector: Relative sign between forward and
backwards propagating part. A plus makes it cosh, a minus makes it sinh.}

\item{deltat}{Numeric: time shift.}
}
\description{
Principal correlator two state model.
}
\seealso{
\code{\link{matrixfit}}
}
