% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hadron-package.R
\docType{data}
\name{plaq.sample}
\alias{plaq.sample}
\title{Sample plaquette time series}
\format{
The format is: num 0.583 0.582 0.582 0.582 0.582 ...
}
\description{
A time series of so-called plaquette values generated by a Markov Chain MC
process using the Hybrid Monte-Carlo algorithm. Plaquettes are the smallest
possible closed loops which can be build in lattice QCD in discretised
Euclidean space time.
}
\examples{

data(plaq.sample)
plot(x=c(1:length(plaq.sample)), y=plaq.sample, type="l", xlab="t", ylab="<P>")

}
\keyword{datasets}
