% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.nlsfit.R
\name{plot.bootstrapfit}
\alias{plot.bootstrapfit}
\title{Plot a bootstrap NLS fit}
\usage{
\method{plot}{bootstrapfit}(x, ..., xlim, ylim, rep = FALSE,
  col.line = "black", col.band = "gray", opacity.band = 0.65,
  lty = c(1), lwd = c(1), xlab = "x", ylab = "y", supports = 1000,
  plot.range, error = sd)
}
\arguments{
\item{x}{object returned by \code{bootstrap.nlsfit}}

\item{...}{Additional parameters passed to the generic \code{plot} function.}

\item{xlim}{x limits of the plot.}

\item{ylim}{y limits of the plot.}

\item{rep}{If set to \code{TRUE}, operate like "replot" in gnuplot. Allows
adding points with error bars to the current plot. Switches the underlying
plotting routine from \code{plot} to \code{points}.}

\item{col.line}{line colour.}

\item{col.band}{error band colour.}

\item{opacity.band}{error band opacity.}

\item{lty}{line type of fitted curve.}

\item{lwd}{line width for fitted curve.}

\item{supports}{number of supporting points for plotting the function.}

\item{plot.range}{vector with two elements \code{c(min,max)} defining the
range in which fitline and errorband are plotted. Default is the range of
the data.}
}
\description{
Plot a bootstrap NLS fit
}
\seealso{
Other NLS fit functions: \code{\link{bootstrap.nlsfit}},
  \code{\link{parametric.bootstrap.cov}},
  \code{\link{parametric.bootstrap}},
  \code{\link{parametric.nlsfit.cov}},
  \code{\link{parametric.nlsfit}},
  \code{\link{print.bootstrapfit}},
  \code{\link{summary.bootstrapfit}}
}
