% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gevp.R
\name{plot.gevp.amplitude}
\alias{plot.gevp.amplitude}
\title{plot.gevp.amplitude}
\usage{
\method{plot}{gevp.amplitude}(x, xlab = "t", ylab = paste0("P[,", x$id, ",", x$op.id, "]"), ...)
}
\arguments{
\item{x}{Object of type \code{gevp.amplitude}.}

\item{xlab}{x axis label}

\item{ylab}{y axis label}

\item{...}{Graphical parameters to be passed on.}
}
\value{
No return value.
}
\description{
plot.gevp.amplitude
}
