% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotutils.R
\name{plot.pionff}
\alias{plot.pionff}
\title{plot.pionff}
\usage{
\method{plot}{pionff}(x, ...)
}
\arguments{
\item{x}{Object of type \code{pionff}}

\item{...}{Generic graphical parameter to be passed on to \link{plotwitherror}}
}
\value{
No return value, only plots are generated.
}
\description{
Generic function to plot an object of type \code{pionff}
}
