% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw.R
\name{plot.rw}
\alias{plot.rw}
\title{Plot a reweighting factor function}
\usage{
\method{plot}{rw}(x, rep = FALSE, ...)
}
\arguments{
\item{x}{\code{rw} object}

\item{rep}{See \code{\link{plotwitherror}}.}
}
\description{
Plot a reweighting factor function
}
