\name{plot.uwerr}
\alias{plot.uwerr}
\title{Plot Command For Class UWerr}
\description{
  Plot Command For Class UWerr
}
\usage{
plot.uwerr <- function(uwerr)
}
\arguments{
  \item{uwerr}{
    object of class \code{uwerr}
  }
}
\value{
  plots
}
\seealso{
  \code{\link{uwerr}}, \code{\link{pion}}, \code{\link{rho}}
}
\author{Carsten Urbach, \email{carsten.urbach@liverpool.ac.uk}}
\keyword{methods}
\keyword{hplot}

