% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.nlsfit.R
\name{print.bootstrapfit}
\alias{print.bootstrapfit}
\title{Print a bootstrap NLS fit}
\usage{
\method{print}{bootstrapfit}(x, digits = 2)
}
\arguments{
\item{x}{object returned by \code{bootstrap.nlsfit}}

\item{digits}{number of significant digits to print in summary or print.}
}
\description{
Print a bootstrap NLS fit
}
\seealso{
Other NLS fit functions: \code{\link{bootstrap.nlsfit}},
  \code{\link{parametric_bootstrap_cov}},
  \code{\link{parametric_bootstrap}},
  \code{\link{parametric_nlsfit_cov}},
  \code{\link{parametric_nlsfit}},
  \code{\link{plot.bootstrapfit}},
  \code{\link{summary.bootstrapfit}}
}
