% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hadron-package.R
\docType{data}
\name{pscor.sample}
\alias{pscor.sample}
\title{Sample pseudoscalar correlator}
\format{
list of 2 elements: "t" "ps"
}
\description{
Sample data for a pseudoscalar correlator for time extent Time=48.
}
\examples{

data("pscor.sample")

}
\keyword{datasets}
