% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed.R
\name{restore_seed}
\alias{restore_seed}
\title{Restore random number generator state}
\usage{
restore_seed(old_seed)
}
\arguments{
\item{old_seed}{integer. Previous seed that should be restored globally.}
}
\value{
No return value, but the random seed is reset to
\code{old_seed}.
}
\description{
Restore random number generator state
}
