% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw.R
\name{rw}
\alias{rw}
\title{Reweighting factor container}
\usage{
rw()
}
\description{
This function \code{rw()} creates containers for reweighting factors
of class \code{rw}. This class is particularly designed to store reweighting
factors for each gauge configuration, that can be applied on
correlation functions emerging in statistical and quantum field theory
simulations. Multiplication operation is defined for this class,
as well as for increasing statistics and \link{is.rw}.
}
\details{
#'
}
\concept{rw constructors}
