% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw.R
\name{rw_meta}
\alias{rw_meta}
\title{rw metadata mixin constructor}
\usage{
rw_meta(.rw = rw(), conf.index = NA)
}
\arguments{
\item{.rw}{\code{rw} object to extend.}

\item{conf.index}{list of Integers, containing the index of gauge configurations}
}
\description{
rw metadata mixin constructor
}
\seealso{
Other cf constructors: \code{\link{cf_boot}},
  \code{\link{cf_meta}}, \code{\link{cf_orig}},
  \code{\link{cf_principal_correlator}},
  \code{\link{cf_shifted}}, \code{\link{cf_smeared}},
  \code{\link{cf_subtracted}}, \code{\link{cf_weighted}},
  \code{\link{cf}}
}
\concept{cf constructors}
