% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf.R
\name{shift.cf}
\alias{shift.cf}
\title{shift a correlation function by 'places' time-slices}
\usage{
shift.cf(cf, places)
}
\arguments{
\item{cf}{unsymmetrised correlation function (cf_meta and cf_orig mixins required)}

\item{places}{integer number of time-slices for backward (negative) or forward (positive) shifts}
}
\value{
Returns an object of class \code{cf} containing the shifted correlation function.
}
\description{
C'(t) = C(t+places)
where places can be positive or negative as required and periodic boundary conditions
in time are assumed
}
