% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_cf.R
\name{/.raw_cf}
\alias{/.raw_cf}
\title{divide two \code{raw_cf} objects}
\usage{
\method{/}{raw_cf}(cf1, cf2)
}
\arguments{
\item{cf1}{'raw_cf' container with data and meta-data to be the dividend}

\item{cf2}{'raw_cf' container with data and meta-data to be the divisor}
}
\value{
\code{raw_cf} object with \code{cf$data == cf1$data / cf2$data}
}
\description{
divide two \code{raw_cf} objects
}
