\name{summary.cfit}
\alias{summary.cfit}
\alias{print.cfit}
\title{Print Summary For Class Cfit}
\description{
  Prints summary for class cfit
}
\usage{
summary(fit)
print(fit)
}
\arguments{
  \item{fit}{
    object of type \code{cfit}
  }
}
\author{Carsten Urbach, \email{carsten.urbach@liverpool.ac.uk}}
\keyword{methods}
\keyword{print}
