% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectivemass.R
\name{summary.effectivemass}
\alias{summary.effectivemass}
\title{summary.effectivemass}
\usage{
\method{summary}{effectivemass}(object, ...)
}
\arguments{
\item{object}{Object of type \code{effectivemass} generated by \link{fit.effectivemass}}

\item{...}{Generic parameters to pass on.}
}
\value{
No return value.
}
\description{
summary.effectivemass
}
