% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf.R
\name{symmetrise.cf}
\alias{symmetrise.cf}
\title{symmetrise.cf}
\usage{
symmetrise.cf(cf, sym.vec = c(1))
}
\arguments{
\item{cf}{Object of type \code{cf}.}

\item{sym.vec}{Integer vector. Takes values -+1 for
(anti)symmmetrisation. If longer than 1, it must be of length
equal to the number of observalbes in \code{cf}. Otherwise, the same
operation is applied to all observables in \code{cf}, which is the
default.}
}
\description{
Symmetrises a correlation function
}
