% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeTemporal.cf.R
\name{takeTimeDiff.cf}
\alias{takeTimeDiff.cf}
\title{Take time difference}
\usage{
takeTimeDiff.cf(cf, deltat = 1, forwardshift = FALSE)
}
\arguments{
\item{cf}{Object of type \code{cf}, a particle correlation function which shall be shifted.}

\item{deltat}{integer. the time shift}

\item{forwardshift}{boolean. If set to \code{TRUE}, the forward finite
difference is used instead of the backward one}
}
\value{
The shifted correlator as an object of type \code{cf}, see \link{cf}
}
\description{
Performs the calculation of the shifted correlator C_shift(t) = C(t) - C(t +/- deltat).
}
