% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_cf.R
\name{uwerr.raw_cf}
\alias{uwerr.raw_cf}
\title{Gamma method analysis on all time-slices in a 'raw_cf' object}
\usage{
uwerr.raw_cf(cf)
}
\arguments{
\item{cf}{Correlation function container of class 'raw_cf'}
}
\value{
The return value is a list with elements
\describe{
\item{\code{value}}{central value}
\item{\code{dvalue}}{statistical error}
\item{\code{ddvalue}}{error of the statistical error}
\item{\code{tauint}}{auto-correlation time estimate}
\item{\code{dtauint}}{error of auto-correlation time estimate}
}
Each of these is in turn an array of dimension \code{ c( cf$nts, cf$dim ) } and
hance lacks the first dimension index compared for \code{cf$data}.
}
\description{
Gamma method analysis on all time-slices in a 'raw_cf' object
}
